==> code/00_run_all.Rmd <==
---
title: "Run all Scripts"
author: Casey Breen
---

## Summary 

This script runs all code to replicate figures and tables. The scripts are numbered in the order they should be run.

Note: The SL3 package is no longer on CRAN, and needs to be installed directly from github. To do so, uncomment and run the following lines: 

```{r}
## install.packages("remotes") 
## remotes::install_github("tlverse/sl3")
```



```{r}
## library packages 
# install.packages("here")
library(here)
source(here("code", "helpers.R"))

## list of Rmd files excluding '00_run_all.Rmd's
rmd_files <- list.files(path = here("code"), pattern = "\\.Rmd$")

## Drop Longer scripts that take a long time to run (best to run individually)
rmd_files <- rmd_files[!rmd_files %in% c("00_run_all.Rmd", "01_fit_models_crossvalidation.Rmd",  "02_make_predictions.Rmd", "10_bootstrap_analysis.Rmd", "12_dhs_table.Rmd")] ##  

## set seed for reproducibility
set.seed(1)

## write output 
con <- file(here("code", "run_all.log"))
sink(con, append=TRUE, split=TRUE)
sink(con, append=TRUE, type = "message")

## Loop to process files
for (cur_file in rmd_files[1:17]) { 

  cat("================================\n")
  source(here("code", "helpers.R"))
  tictoc::tic(glue::glue("Finished with {cur_file}, time:"))
  cat("Running ", cur_file, "\n")
  source_rmd(here("code", cur_file))
  tictoc::toc()
  cat("================================\n")                                                                                             

}
sink()
sink()
```
