==> code/01_fit_models_crossvalidation.Rmd <==
---
title: "Fit models using cross-validation"
author: Casey Breen
---

Summary: Fit subnational models using both leave-one-country-out (LOCO) and 10-fold cross-validation.


## Library packages 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```

## Read in data 

```{r}
## read in training and test datasets 
master_file <- read_csv(here("data","master_data_file_march12_with_national.csv")) %>% 
  filter(!(country ==  "Tanzania" & dhsyear == 2015)) %>% ## drop earlier surveys and Lesotho (only became available after R&R)
  filter(!(country ==  "Nepal" & dhsyear == 2016)) %>% 
  filter(!(country == "Lesotho"))

## read in covariates 
master_file <- master_file %>% 
  dplyr::select(country, gid_0, gid_1, year, perc_used_internet_past12months_wght_age_15_to_49_wom, continent, all_of(c(dhs_vars, fb_feats, off_feats))) %>% ## the vectors of vars are defined in helper function
  mutate(continent = case_when(continent == "Africa" ~ 1,
                               TRUE ~ 0)) %>%
  drop_na(dhs_vars, fb_feats) %>% 
  mutate(perc_used_internet_past12months_wght_age_15_to_49_fm_ratio =  perc_used_internet_past12months_wght_age_15_to_49_wom/perc_used_internet_past12months_wght_age_15_to_49_men,
         perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio =  perc_owns_mobile_telephone_wght_age_15_to_49_wom/perc_owns_mobile_telephone_wght_age_15_to_49_men)
```


```{r}
## check missing
master_file %>%
  summarise(across(everything(), ~sum(is.na(.)))) %>%
  pivot_longer(-1) 

## impute missing values as median (only necessary for a handful of cases)
master_file <- master_file %>% 
  mutate(across(where(is.numeric), ~ ifelse(is.na(.), median(., na.rm = TRUE), .)))

## top code ratio variables 
master_file <- master_file %>%
  mutate(perc_used_internet_past12months_wght_age_15_to_49_fm_ratio = case_when(
    perc_used_internet_past12months_wght_age_15_to_49_fm_ratio > 1.5 ~ 1.5,
    TRUE ~ perc_used_internet_past12months_wght_age_15_to_49_fm_ratio
  )) %>%
  mutate(perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio = case_when(
    perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio > 1.5 ~ 1,
    TRUE ~ perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio
  ))

## check missing values
master_file <- master_file %>% 
  mutate(dhsyear = case_when(
  country == "India" & dhsyear == 2019 ~ 2020,
  TRUE ~ dhsyear)) 

master_file %>%
  group_by(country) %>%
  filter(!is.na(perc_used_internet_past12months_wght_age_15_to_49_wom)) %>%
  filter(!is.na(fb_pntr_18p_female)) %>%
  summarize(cor = cor(perc_used_internet_past12months_wght_age_15_to_49_wom, fb_pntr_18p_female),
            n()) %>% 
  arrange(cor) 
```

## make predictions 

```{r}
# Define a list of outcomes and corresponding feature sets
outcomes <- c(
  "perc_used_internet_past12months_wght_age_15_to_49_wom", 
   "perc_owns_mobile_telephone_wght_age_15_to_49_wom",
   "perc_owns_mobile_telephone_wght_age_15_to_49_men",
   "perc_used_internet_past12months_wght_age_15_to_49_men",
   "perc_used_internet_past12months_wght_age_15_to_49_fm_ratio",
   "perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio"
 )

## define feature sets 
feature_sets <- list(
  fb = fb_feats,
  offline = off_feats,
  fb_and_offline = c(fb_feats, off_feats) # , national_covars national_covars
)

# Function to make predictions based on model type
make_predictions <- function(data, outcomes, feature_sets, model_function) {
  prediction_list <- list()

  for (outcome in outcomes) {
    for (feature_set_name in names(feature_sets)) {
      prediction_name <- paste0("predictions_", outcome, "_", feature_set_name)
      prediction_list[[prediction_name]] <- model_function(data, outcome, feature_sets[[feature_set_name]])
      cat(prediction_name, "\n")
    }
  }
  

  # Combine all prediction results into a single dataframe
  bind_rows(prediction_list) 
}

# test <- run_model_10fold(data = gadm1, outcome = "perc_used_internet_past12months_wght_age_15_to_49_wom", covar_list = c(off_feats))

## set seed for reproducibility
set.seed(3)

# Predictions using LOCO model
predictions_loco <- make_predictions(master_file, outcomes, feature_sets, run_loco_model) %>%  
    mutate(cv_method = "loco")

# Predictions using 10-fold model
predictions_10fold <- make_predictions(master_file, outcomes, feature_sets, run_model_10fold) %>%
    mutate(cv_method = "10fold")

## combine predictions 
predictions_combined <- bind_rows(predictions_loco, predictions_10fold)

predictions_combined <- predictions_combined %>% 
    mutate(outcome_clean = case_when(
    grepl("mobile_telephone_wght_age_15_to_49_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_telephone_wght_age_15_to_49_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_telephone_wght_age_15_to_49_wom", outcome) ~ "Mobile (Women)",
    grepl("used_internet_past12months_wght_age_15_to_49_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("used_internet_past12months_wght_age_15_to_49_men", outcome) ~ "Internet (Men)",
    grepl("used_internet_past12months_wght_age_15_to_49_wom", outcome) ~ "Internet (Women)",
    TRUE ~ outcome
  )) 
```


```{r}
# Write the LOCO predictions to a .csv file
write_csv(predictions_combined, here("out", "combined_predictions_lmics_national_covars_full.csv"))
```
