==> code/03_model_performance.Rmd <==
---
title: "Model Peformance"
auhor: "Casey Breen" 
---

Summary: Generate figures and tables summarizing model performance 
 
```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```

## Estimate model performance metrics 

```{r}
## read predictions
predictions <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>%
  filter(country != "Lesotho") %>% # & country != "Gambia" 
 # filter(!gid_1 %in% predictions_old$gid_1) %>% 
mutate(predicted = case_when(
    predicted < 0 ~ 0,
    predicted > 1 ~ 1,
    TRUE ~ predicted
  )) %>%
  mutate(observed = case_when(
    observed < 0 ~ 0,
    observed > 1 ~ 1,
    TRUE ~ observed
  ))  %>% 
  mutate(cv_method = case_when(
    grepl("10fold", cv_method) ~ "10-Fold-CV",
    grepl("loco", cv_method) ~ "LOCO-CV",
    TRUE ~ cv_method
  )) %>%
  mutate(outcome_categorical = case_when(
    outcome %in% c(
      "perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio",
      "perc_owns_mobile_telephone_wght_age_15_to_49_men",
      "perc_owns_mobile_telephone_wght_age_15_to_49_wom"
    ) ~ "Mobile",
    outcome %in% c(
      "perc_used_internet_past12months_wght_age_15_to_49_fm_ratio",
      "perc_used_internet_past12months_wght_age_15_to_49_men",
      "perc_used_internet_past12months_wght_age_15_to_49_wom"
    ) ~ "Internet",
  )) 


## read in training and test datasets
gadm1 <- read_csv(here("data", "master_data_file_march12_with_national.csv"),
  locale = locale(encoding = "ISO-8859-1")) %>%
  janitor::clean_names() %>%
  mutate(pop_all_age_total_dhsyear = pop_all_age_female + pop_all_age_male) %>%
  dplyr::select(country, gid_0, gid_1, dhsyear, pop_all_age_total_dhsyear, continent) %>%
  filter(!is.na(dhsyear)) 


predictions_combined <- predictions %>%
  select(-country) %>% 
  left_join(gadm1 %>% 
              mutate(country = case_when(
                gid_0 == "CIV" ~ "Côte d'Ivoire", TRUE ~ country)), by = c("gid_0", "gid_1", "year" = "dhsyear")) 

## estimate model performance metrics
## r2, rmse, mae are fine
## mape and mare are a bit tricky because observed value is sometimes 0
model_performance_metrics <- predictions_combined %>%
  group_by(cv_method, model, outcome, feature_set) %>%
  summarise(
    r = cor(predicted, observed),
    r2 = compute_r2(predicted, observed),
    rmse = compute_rmse(predicted, observed),
    mape = compute_mape(predicted, observed),
    mare = compute_mare(predicted, observed),
    mae = compute_mae(predicted, observed)
  ) %>%
  pivot_longer(
    cols = c(r, r2, rmse, mape, mare, mae),
    names_to = "metric",
    values_to = "value"
  )

## model performance metrics 
model_performance_metrics <- model_performance_metrics %>%
  mutate(outcome_clean = case_when(
    grepl("mobile_telephone_wght_age_15_to_49_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_telephone_wght_age_15_to_49_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_telephone_wght_age_15_to_49_wom", outcome) ~ "Mobile (Women)",
    grepl("used_internet_past12months_wght_age_15_to_49_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("used_internet_past12months_wght_age_15_to_49_men", outcome) ~ "Internet (Men)",
    grepl("used_internet_past12months_wght_age_15_to_49_wom", outcome) ~ "Internet (Women)",
    TRUE ~ outcome
  )) 

## model performance 
model_performance_metrics %>% filter(metric == "r2")

## predictions combined 
predictions_combined <- predictions_combined %>%
  left_join(model_performance_metrics %>% filter(metric == "r2") %>% mutate(r2 = value) %>% select("r2", "metric", "cv_method", "model", "outcome", "feature_set"), by = c("cv_method", "model", "outcome", "feature_set")) %>%
    left_join(model_performance_metrics %>% filter(metric == "r") %>% mutate(r = value) %>% select("r", "metric", "cv_method", "model", "outcome", "feature_set"), by = c("cv_method", "model", "outcome", "feature_set")) %>% 
    left_join(model_performance_metrics %>% filter(metric == "mae") %>% mutate(mae = value) %>% select("mae", "metric", "cv_method", "model", "outcome", "feature_set"), by = c("cv_method", "model", "outcome", "feature_set")) 
```

## observed vs. predicted by outcome 

```{r}
# Plot for "perc_ever_used_internet_wght_age_15_to_49_men"
plot1_observed_predicted <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_wom", "perc_used_internet_past12months_wght_age_15_to_49_wom"),
    feature_set == "fb_and_offline",
    cv_method == "LOCO-CV"
  ) %>%
  ggplot(aes(x = observed, y = predicted)) + # size = pop_all_age_total_dhsyear
  geom_point(alpha = 0.1) +
  geom_abline(slope = 1, linetype = "dashed", color = "grey") +
  theme_cowplot() +
  labs(title = "Women") +
  geom_label(
  aes(
    label = paste0(
  "R\u00B2 = ", round(r2, 2), "\n",
  "r = ", round(r, 2), "\n",
  "MAE = ", round(mae, 2))), 
  x = 0.20,
  y = 0.90,
  size = 3) +
  scale_y_continuous(limits = c(0, 1.1)) +
  scale_x_continuous(limits = c(0, 1.1)) +
  facet_wrap(~outcome_categorical, nrow = 2) +
  labs(x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)") + 
   theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.3, "lines"),
    aspect.ratio = .8,
    legend.position = "none",
    plot.margin = margin(0,0,0,0)
   )

# Plot for "perc_ever_used_internet_wght_age_15_to_49_"
plot2_observed_predicted <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_men", "perc_used_internet_past12months_wght_age_15_to_49_men"),
    feature_set == "fb_and_offline",
    cv_method == "LOCO-CV"
  ) %>%
  ggplot(aes(x = observed, y = predicted, size = pop_all_age_total_dhsyear)) +
  geom_point(alpha = 0.1) +
    scale_y_continuous(limits = c(0, 1.1)) +
    scale_x_continuous(limits = c(0, 1.1)) +
  geom_abline(slope = 1, linetype = "dashed", color = "grey") +
  theme_cowplot() +
  labs(title = "Men") +
   geom_label(
  aes(
    label = paste0(
  "R\u00B2 = ", round(r2, 2), "\n",
  "r = ", round(r, 2), "\n",
  "MAE = ", round(mae, 2))), 
  x = 0.20,
  y = 0.90,
  size = 3) +
  facet_wrap(~outcome_categorical, nrow = 2) +
  labs(x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)") + 
   theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.3, "lines"),
    aspect.ratio = .8,
    legend.position = "none",
    plot.margin = margin(0,0,0,0)
  ) 

# Plot for "perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio"
plot3_observed_predicted <- predictions_combined %>%
  # mutate(r2 = paste0("italic(R)^2~`=`~", round(r2, 2))) %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio", "perc_used_internet_past12months_wght_age_15_to_49_fm_ratio"),
    feature_set == "fb_and_offline",
    cv_method == "LOCO-CV"
  ) %>%
  ggplot(aes(x = observed, y = predicted, size = pop_all_age_total_dhsyear)) +
  geom_point(alpha = 0.1) +
  ylim(0, 1.1) +
  xlim(0, 1.1) +
  geom_abline(slope = 1, linetype = "dashed", color = "darkgrey") +
  theme_cowplot() +
  labs(title = "Gender Gap Index", size = "Population Size") +
     geom_label(aes(
    label = paste0(
  "R\u00B2 = ", round(r2, 2), "\n",
  "r = ", round(r, 2), "\n",
  "MAE = ", round(mae, 2))), 
  x = 0.20,
  y = 0.90,
  size = 3) +
  facet_wrap(~outcome_categorical, nrow = 2) +
  theme(legend.position = "none") +
  scale_size_continuous(
    breaks = c(300000, 1000000, 3000000, 10000000),
    labels = c("300k", "1m", "3m", "10m")
  ) +
  guides(size = guide_legend(override.aes = list(alpha = 1))) + 
  labs(x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)") + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.3, "lines"),
    aspect.ratio = .8,
    legend.position = "none",
    plot.margin = margin(0,0,0,0)
  ) 

# Combine plots using cowplot's plot_grid
combined_plot_observed_predicted <- plot_grid(plot1_observed_predicted, plot2_observed_predicted, plot3_observed_predicted, ncol = 3, labels = "AUTO",   rel_widths = c(1, 1, 1)  # optional but keeps balance
)

# Extract the legend from plot3_internet12m (or from whichever plot you prefer)
legend <- ggpubr::get_legend(
  plot3_observed_predicted + theme(legend.position = "bottom", legend.direction = "horizontal")
)

# Combine the combined_plots with the extracted legend and place it at the bottom
combined_plot_observed_predicted_legend <- plot_grid(combined_plot_observed_predicted, legend, ncol = 1, rel_heights = c(1, .1))

ggsave(plot = combined_plot_observed_predicted_legend, here("figures", "observed_predicted_loco_complete_fig.png"), width = 10, height = 6)
ggsave(plot = combined_plot_observed_predicted_legend, here("figures", "fig2_observed_predicted_loco.pdf"), width = 10, height = 6)

```



```{r}
plot1_combined <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_wom", "perc_used_internet_past12months_wght_age_15_to_49_wom"),
    feature_set == "fb_and_offline") %>% 
  ggplot(aes(x = observed, y = predicted)) +
  geom_point(aes(size = pop_all_age_total_dhsyear), alpha = 0.1) +
  # geom_smooth() +
    scale_y_continuous(limits = c(0, 1.1)) +
    scale_x_continuous(limits = c(0, 1.1)) +
  geom_abline(slope = 1, linetype = "dashed", color = "grey") +
  theme_cowplot() +
  geom_label(
  aes(
    label = paste0(
  "R\u00B2 = ", round(r2, 2), "\n",
  "r = ", round(r, 2), "\n",
  "MAE = ", round(mae, 2))), 
  x = 0.20,
  y = 0.90,
  size = 3.5) +
  facet_grid(rows = vars(outcome_categorical), cols = vars(cv_method)) +
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.3, "lines"),
    aspect.ratio = .8,
    legend.position = "none",
    plot.margin = margin(0,0,0,0)
  )  + 
  labs(title = "Women",
       x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)")


# Plot for "perc_ever_used_internet_wght_age_15_to_49_wom"
plot2_combined <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_men", "perc_used_internet_past12months_wght_age_15_to_49_men"),
    feature_set == "fb_and_offline"
  ) %>%
  ggplot(aes(x = observed, y = predicted)) +
  geom_point(aes(size = pop_all_age_total_dhsyear), alpha = 0.1) +
  # geom_smooth() +
  scale_y_continuous(limits = c(0, 1.1)) +
  scale_x_continuous(limits = c(0, 1.1)) +
  geom_abline(slope = 1, linetype = "dashed", color = "grey") +
  theme_cowplot() +
  geom_label(
    aes(
      label = paste0(
        "R\u00B2 = ", round(r2, 2), "\n",
        "r = ", round(r, 2), "\n",
        "MAE = ", round(mae, 2))), 
    x = 0.20,
    y = 0.90,
    size = 3.5) +
  facet_grid(rows = vars(outcome_categorical), cols = vars(cv_method)) +
  theme(legend.position = "none") + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.3, "lines"),
    aspect.ratio = .8,
    legend.position = "none",
    plot.margin = margin(0,0,0,0)
  ) + 
  labs(title = "Men",
       x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)")


plot3_combined <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio", "perc_used_internet_past12months_wght_age_15_to_49_fm_ratio"),
    feature_set == "fb_and_offline"
  ) %>%
  ggplot(aes(x = observed, y = predicted)) +
  geom_point(aes(size = pop_all_age_total_dhsyear), alpha = 0.1) +
  ylim(0, 1.1) +
  xlim(0, 1.1) +
  geom_abline(slope = 1, linetype = "dashed", color = "grey") +
  theme_cowplot() +
  geom_label(
    aes(
      label = paste0(
        "R\u00B2 = ", round(r2, 2), "\n",
        "r = ", round(r, 2), "\n",
        "MAE = ", round(mae, 2))), 
    x = 0.2,
    y = .90,
    size = 3.5) +
  facet_grid(rows = vars(outcome_categorical), cols = vars(cv_method)) +
  theme(legend.position = "none") +
  scale_size_continuous(
    breaks = c(300000, 1000000, 3000000),
    labels = c("300k", "1m", "3m")
  ) +
  guides(size = guide_legend(override.aes = list(alpha = 1))) + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.3, "lines"),
    aspect.ratio = .8,
    legend.position = "none",
    plot.margin = margin(0,0,0,0)
  ) + 
  labs(x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)",
       title = "Gender Gap",
       size = "Population Size")


# Extract the legend from plot3_internet12m (or from whichever plot you prefer)
legend <- cowplot::get_legend(
  plot3_combined + theme(legend.position = "bottom", legend.direction = "horizontal")
)

# Combine the plots without legends
plot_combined_observed_predicted <- plot_grid(plot1_combined, plot2_combined, plot3_combined, ncol = 1, labels = "AUTO", align = "v")

# Combine the combined_plots with the extracted legend and place it at the bottom
plot_combined_observed_predicted_legend <- plot_grid(plot_combined_observed_predicted, legend, ncol = 1, rel_heights = c(1, .1))

ggsave(plot = plot_combined_observed_predicted_legend, here("figures" ,"plot_combined_observed_predicted.png"), width = 6, height = 18)
```



## Observed vs. predicted by country 

```{r}
# Define outcomes and labels
outcomes <- list(
  perc_owns_mobile_telephone_wght_age_15_to_49_wom = "Own mobile, women",
  perc_owns_mobile_telephone_wght_age_15_to_49_men = "Own mobile, men",
  perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio = "Own mobile, gender gap",
  perc_used_internet_past12months_wght_age_15_to_49_wom = "Internet, women",
  perc_used_internet_past12months_wght_age_15_to_49_men = "Internet, men",
  perc_used_internet_past12months_wght_age_15_to_49_fm_ratio = "Internet, gender gap"
)

predictions_combined <- predictions %>%
  left_join(gadm1 %>% 
              mutate(country = case_when(
                gid_0 == "CIV" ~ "Côte d'Ivoire", TRUE ~ country)), by = c("gid_0", "gid_1", "year" = "dhsyear", "country")) 

# Precompute metrics
metrics_by_country_outcome <- predictions_combined %>%
  filter(
    model == "superlearner",
    cv_method == "LOCO-CV",
    feature_set == "fb_and_offline"
  ) %>%
  group_by(outcome, gid_0) %>%
  summarize(
    r = cor(observed, predicted, use = "complete.obs"),
    mae = mean(abs(observed - predicted), na.rm = TRUE),
    .groups = "drop"
  )

# Loop over outcomes
for (i in seq_along(outcomes)) {
  outcome_var <- names(outcomes)[i]
  title_text <- outcomes[[i]]
  
  # Filter and join metrics
  df <- predictions_combined %>%
    filter(
      model == "superlearner",
      cv_method == "LOCO-CV",
      feature_set == "fb_and_offline",
      outcome == outcome_var
    ) %>%
    left_join(
      metrics_by_country_outcome %>% filter(outcome == outcome_var),
  by = c("gid_0", "outcome")
    )
  
  # Create plot
  plot_obj <- ggplot(df, aes(x = observed, y = predicted)) +
    geom_point(alpha = 0.3, aes(size = pop_all_age_total_dhsyear)) +
    geom_abline(slope = 1, linetype = "dashed") + 
    facet_wrap(~country) +
    scale_size_continuous(
      breaks = c(300000, 1000000, 3000000, 10000000),
      labels = c("300k", "1m", "3m", "10m")
    ) +
    guides(size = guide_legend(override.aes = list(alpha = 1))) +
    geom_label(
    aes(
      label = paste0("atop(italic(r)==", round(r, 2), ",~italic(MAE)==", round(mae, 2), ")")
    ),
    parse = TRUE,
    x = 0.25,
    y = 0.9,
    size = 2,
    inherit.aes = FALSE
  ) +
    theme_cowplot() +
    xlim(0, 1.1) + 
    ylim(0, 1.1) +
    theme(
      strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
      strip.text = element_text(size = 9, face = "bold", color = "black"),
      panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
      panel.background = element_rect(fill = "#f9f9f9", color = NA),
      panel.grid = element_line(color = "grey95"),
      panel.spacing = unit(.2, "lines"),
      aspect.ratio = .9,
      axis.text = element_text(size = 11),            # Axis tick labels
      legend.position = "bottom") + 
    labs(x = "Observed (Admin-1)", y = "Predicted (Admin-1)", size = "Population Size")
  
  # Save plot
  ggsave(
    plot = plot_obj,
    filename = here("figures", paste0("observed_predicted_plot_country_", outcome_var, ".png")),
    width = 10,
    height = 9
  )
}
```





```{r}
country_r2_df <- predictions_combined %>%
  group_by(cv_method, model, outcome, feature_set, country) %>%
    filter(
    model == "superlearner",
    feature_set == "fb_and_offline"
  ) %>%
  summarise(
    r2 = compute_r2(predicted, observed),
    r2_alt = calculate_r_squared(predicted, observed),
    cor = cor(predicted, observed),
    mae = mean(abs(observed - predicted)),
    mse = mean(sqrt((observed - predicted)^2))
  )

country_level_mae <- country_r2_df %>%
  filter(outcome == "perc_used_internet_past12months_wght_age_15_to_49_wom") %>%
  ggplot(aes(x = mae)) +
  geom_density(aes(color = cv_method, fill = cv_method, y = after_stat(count) / sum(after_stat(count))), alpha = 0.2) +
  labs(x = "Mean absolute error", y = "Proportion") +
  theme_cowplot() +
 # xlim(0, 1) +
  theme(legend.position = "bottom", legend.title = element_blank()) +
  scale_fill_manual(values = cud[1:2]) + 
  scale_color_manual(values = cud[1:2]) 

country_level_corr <- country_r2_df %>%
  filter(outcome == "perc_used_internet_past12months_wght_age_15_to_49_wom") %>%
  ggplot(aes(x = cor)) +
  geom_density(aes(color = cv_method, fill = cv_method, y = after_stat(count) / sum(after_stat(count))), alpha = 0.2) +
  labs(x = "Correlation", y = "Proportion") +
  theme_cowplot() +
 # xlim(0, 1) +
  theme(legend.position = "bottom", legend.title = element_blank()) +
  scale_fill_manual(values = cud[1:2]) + 
  scale_color_manual(values = cud[1:2]) + 
  xlim(0, 1)

## save plots as rds bc we'll need to use them later in a multi-panel plot 
write_rds(country_level_mae, file = here("figures", "country_level_mae.rds"))
write_rds(country_level_corr, file = here("figures", "country_level_corr.rds"))

```

```{r}
# Combine `outcome_categorical` and `dhsyear` into a single facet variable
predictions_combined <- predictions_combined %>%
  mutate(facet_variable = paste(outcome_categorical, year, sep = " - "))

# Compute MAE and correlation
predictions_combined_summary <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_used_internet_past12months_wght_age_15_to_49_wom"),
    feature_set == "fb_and_offline",
    cv_method == "LOCO-CV"
  ) %>%
  group_by(facet_variable) %>%
  summarise(
    mae = mean(abs(predicted - observed)),
    r = cor(observed, predicted)
  )

# Plot
loco_by_year_fig <- predictions_combined %>%
  filter(
    model == "superlearner",
    outcome %in% c("perc_used_internet_past12months_wght_age_15_to_49_wom"),
    feature_set == "fb_and_offline",
    cv_method == "LOCO-CV"
  ) %>%
  ggplot(aes(x = observed, y = predicted, size = pop_all_age_total_dhsyear)) +
  geom_point(alpha = 0.2) +
  geom_abline(slope = 1, linetype = "dashed", color = "grey") +
  theme_cowplot() +
  labs(title = "Women",
       size = "Population size") +
  scale_y_continuous(limits = c(0, 1.1)) +
  scale_x_continuous(limits = c(0, 1.1)) +
  facet_wrap(~facet_variable, nrow = 2) +
  theme(legend.position = "bottom") +
  geom_label(
    data = predictions_combined_summary,
    aes(
      x = 0.3, 
      y = 0.9,
      label = paste("atop(italic(r) == ", round(r, 2), ", italic(MAE) == ", round(mae, 2), ")")
    ),
    parse = TRUE,
    inherit.aes = FALSE,
    size = 3
  ) + 
   scale_size_continuous(
    breaks = c(300000, 1000000, 3000000, 10000000),
    labels = c("300k", "1m", "3m", "10m")
  ) +
  guides(size = guide_legend(override.aes = list(alpha = 1))) + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(.5, "lines"),
    aspect.ratio = .9,
    legend.position = "bottom",
    legend.justification = "center",  # centers the whole legend box
    legend.title = element_text(size = 13, hjust = 0.5, vjust = 1, margin = margin(b = 4)),
    legend.text = element_text(size = 11)
  ) + 
  labs(x = "Observed (Admin-1)",
       y = "Predicted (Admin-1)")

ggsave(plot = loco_by_year_fig, filename = here("figures", "loco_by_year_fig.png"), width = 9, height = 5)
```





```{r}
df_for_plot <- predictions_combined %>%
  filter(
    model == "superlearner",
    feature_set == "fb_and_offline",
    cv_method == "LOCO-CV"
  ) %>% 
  group_by(outcome_clean, year) %>% 
  summarize(error = mean(observed - predicted),
            error_absolute = mean(abs(observed - predicted))) 

error_over_time_part1 <- df_for_plot %>% 
  ggplot(aes(x = year, y = error, color = outcome_clean)) + 
  geom_line() + 
  geom_point(fill= "white", shape = 21, size = 2) + 
  theme_cowplot() + 
  ylim(-0.35, 0.35) + 
  geom_hline(yintercept = mean(df_for_plot$error), linetype = "dashed", color = "black") + 
  theme_cowplot() + 
  scale_color_manual(values = cbp1) + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Year",
       y = "Mean Error \n (Observed - Predicted)")


error_over_time_part2 <- df_for_plot %>% 
  ggplot(aes(x = year, y = error_absolute, color = outcome_clean)) + 
  geom_line() + 
  geom_point(fill= "white", shape = 21, size = 2) + 
  theme_cowplot() + 
  ylim(0, .5) + 
  geom_hline(yintercept = mean(df_for_plot$error_absolute), linetype = "dashed", color = "black") + 
  theme_cowplot() + 
  scale_color_manual(values = cbp1) + 
  theme(legend.position = "bottom", legend.title = element_blank()) + 
  labs(x = "Year",
       y = "Mean Absolute Error")

combined_fig <- plot_grid(error_over_time_part1, error_over_time_part2, nrow = 2, labels = "AUTO")

ggsave(plot = combined_fig, filename = here("figures", "error_over_time.png"), height = 8, width = 6)
```
