==> code/04_within_country_analysis.Rmd <==
---
title: "Within-country variation"
author: Casey Breen
---

Summary: Create plots of within-country variation in internet adoption and mobile ownership. 


```{r}
## library packages + custom functions 
library(here)
source(here("code", "helpers.R"))

## read in prediction 
predictions_long <- read_csv(here("out", "predictions_with_uncertainty.csv"))

predictions_wide <- read_csv(here("out", "predictions_wide.csv"))

## predictions continent 
predictions_wide$continent <- countrycode(predictions_wide$gid_0, origin = 'iso3c', destination = 'continent')

## predictions continent 
predictions_long$continent <- countrycode(predictions_long$gid_0, origin = 'iso3c', destination = 'continent')
```


```{r}
spread_plot_df <- predictions_long %>% 
  filter(date == "2025-01") %>% 
  mutate(outcome_clean = case_when(
    grepl("mobile_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_women", outcome) ~ "Mobile (Women)",
    grepl("internet_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("internet_men", outcome) ~ "Internet (Men)",
    grepl("internet_women", outcome) ~ "Internet (Women)",
    TRUE ~ outcome
  )) %>% 
  # filter(!gid_1 %in% c("MRT.9_1", "UGA.32_1", "UGA.31_1", "MKD.1_1", "MKD.10_1")) %>% 
  group_by(continent, gid_0, outcome_clean) %>% 
  summarize(diff = max(predicted) - min(predicted),
            mean = mean(predicted))
  

# Define the order of the levels explicitly
spread_plot_df <- spread_plot_df %>% 
  mutate(outcome_clean = factor(outcome_clean, 
                                levels = c("Internet (Women)", "Internet (Men)", "Internet (Ratio)",
                                           "Mobile (Women)", "Mobile (Men)", "Mobile (Ratio)")))

spread_plot_facet <- spread_plot_df %>% 
  ggplot(aes(x = mean, y = diff)) + 
  geom_text(aes(label = gid_0, color = continent)) + 
  labs(x = "National Estimate",
       y = "Within-Country Range",
       color = "Region") + 
  theme_cowplot() + 
  scale_color_manual(values = cud) + 
  geom_smooth(se = F, method = "lm") + 
  theme(legend.position = "bottom") + 
  stat_cor(method = "pearson", label.x = Inf, label.y = Inf, size = 5.5,  hjust=1.1, vjust = 4.4, p.accuracy = 0.01,
            aes(label = paste0("italic(r)~`=`~", ..r..))) + 
  facet_wrap(~outcome_clean, scales = "free") + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(1.5, "lines"),
    aspect.ratio = .9,
    legend.position = "bottom",
    legend.justification = "center",  # centers the whole legend box
    legend.title = element_text(size = 13, hjust = 0.5, vjust = 1, margin = margin(b = 4)),
    legend.text = element_text(size = 11)
  ) 

ggsave(spread_plot_facet, filename = here("figures", "spread_plot_facet.png"), width = 12, height = 8)
```



```{r}
## create spread plot 
spread_plot <- predictions_long %>% 
    filter(date == "2025-01") %>% 
  filter(outcome == "internet_women") %>% 
  group_by(continent, gid_0) %>% 
  summarize(diff = max(predicted) - min(predicted),
            mean = mean(predicted)) %>% 
  ggplot(aes(x = mean, y = diff)) + 
  geom_text(aes(label = gid_0, color = continent)) + 
  labs(x = "National Internet Adoption (Women)",
       y = "Within-Country Range (Women)",
       color = "Region") + 
  theme_cowplot() + 
  scale_color_manual(values = cud) + 
  geom_smooth(se = F, method = "lm") + 
  theme(legend.position = "bottom") + 
  stat_cor(label.x = .65, label.y = .55, size = 6, p.accuracy = 0.01, aes(label = paste0("italic(r)~`=`~", ..r..)))

## write rds 
write_rds(spread_plot, file = here("figures", "spread_plot.rds"))
```




```{r}
predictions_wide <- predictions_wide %>% 
  filter(year == 2024 & month == 11) %>% 
  mutate(across(c(predicted_perc_used_internet_past12months_wght_age_15_to_49_wom, predicted_perc_owns_mobile_telephone_wght_age_15_to_49_wom, predicted_perc_used_internet_past12months_wght_age_15_to_49_men, predicted_perc_owns_mobile_telephone_wght_age_15_to_49_men), ~ ifelse(. > 1, 1, .))) 

internet_vs_mobile_predictions_women <- predictions_wide %>% 
  ggplot(aes(x = predicted_perc_used_internet_past12months_wght_age_15_to_49_wom, y =predicted_perc_owns_mobile_telephone_wght_age_15_to_49_wom )) + 
  geom_point(aes(color = continent), alpha = 0.4) + 
  theme_cowplot() + 
  scale_color_manual(values = cud) + 
  labs(x = "Predicted Internet (Women)",
       y = "Predicted Mobile Ownership (Women)",
       fill = "Gender Ratio") + 
    ylim(0, 1.2) + 
  xlim(0, 1.2) + 
  theme(legend.position = "none")

internet_vs_mobile_predictions_men <- predictions_wide %>% 
  ggplot(aes(x = predicted_perc_used_internet_past12months_wght_age_15_to_49_men, y =predicted_perc_owns_mobile_telephone_wght_age_15_to_49_men )) + 
  geom_point(aes(color = continent), alpha = 0.4) + 
  theme_cowplot() + 
  scale_color_manual(values = cud) + 
  labs(x = "Predicted Internet (Men)",
       y = "Predicted Mobile Ownership (Men)",
       fill = "Gender Ratio") + 
  ylim(0, 1.2) + 
  xlim(0, 1.2) + 
  theme(legend.position = "none")

internet_vs_mobile_predictions_fm_ratio <- predictions_wide %>% 
  ggplot(aes(x = predicted_perc_used_internet_past12months_wght_age_15_to_49_fm_ratio, y =predicted_perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio )) + 
  geom_point(aes(color = continent), alpha = 0.4) + 
  theme_cowplot() + 
  labs(x = "Predicted Internet (Ratio)",
       y = "Predicted Mobile Ownership (Ratio)",
       fill = "Gender Ratio") + 
  ylim(0, 1.5) + 
  xlim(0, 1.5) + 
  scale_color_manual(values = cud) + 
  theme(legend.position = c(0.95, 0.05), 
        legend.justification = c("right", "bottom"),
        legend.box = "vertical",
        legend.title = element_blank(),
        legend.background = element_rect(color = "black", linewidth = 0.5),
        legend.key.size = unit(0.5, "cm")) 

combined_internet_mobile_plot <- cowplot::plot_grid(internet_vs_mobile_predictions_women, 
                   internet_vs_mobile_predictions_men,
                   internet_vs_mobile_predictions_fm_ratio,
                   nrow = 3,
                   labels = 'AUTO')

ggsave(combined_internet_mobile_plot, filename = here("figures", "combined_internet_mobile_plot.png"), width = 4.7, height = 13)
```
