==> code/05_geo_map.Rmd <==
---
title: "Create Maps"
author: Casey Breen
---

Summary: Create maps and other visualizations 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```

```{r}
## read in training and test datasets 
master_file <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>%
  dplyr::select(country, all_of(c(dhs_vars, fb_feats, off_feats))) %>% 
  filter(!is.na(perc_used_internet_past12months_wght_age_15_to_49_wom)) 

master_file_full <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>% 
  filter(!is.na(fb_pntr_18p_female)) %>% 
  filter(year == 2025 & month == 01)


# Define the list of variables
variables <- c(
  "predicted_perc_owns_mobile_telephone_wght_age_15_to_49_wom",
  "predicted_perc_owns_mobile_telephone_wght_age_15_to_49_men",
  "predicted_perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio",
  "predicted_perc_used_internet_past12months_wght_age_15_to_49_wom",
  "predicted_perc_used_internet_past12months_wght_age_15_to_49_men",
  "predicted_perc_used_internet_past12months_wght_age_15_to_49_fm_ratio"
)

## read in predictions 
predictions_wide <- read_csv(here("out", "predictions_wide.csv")) %>% 
    mutate(across(all_of(variables), ~ case_when(. > 1 ~ 1, TRUE ~ .)))

## predictions with uncertainty 
predictions_uncertainty <- read_csv(here("out", "predictions_with_uncertainty.csv")) %>% 
  filter(outcome == "internet_women") %>% 
  filter(date == "2025-01") 

## loco predictions 
loco_predictions <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>% 
  filter(model == "superlearner" & feature_set == "fb_and_offline" & cv_method == "loco") 
```


```{r}
st_layers(here("data", "gadm_410-levels.gpkg"))
gadm1_map <- read_sf(here("data", "gadm_410-levels.gpkg"), layer = "ADM_1")
gadm0_map <- read_sf(here("data", "gadm_410-levels.gpkg"), layer = "ADM_0")

gadm0_map <- gadm0_map %>% 
  filter(COUNTRY %in% master_file_full$country | COUNTRY == "Côte d'Ivoire" | COUNTRY == "Sudan" | COUNTRY == "South Sudan")

gadm1_map <- gadm1_map %>% 
  filter(COUNTRY %in% master_file_full$country | COUNTRY == "Côte d'Ivoire" | COUNTRY == "Sudan"| COUNTRY == "South Sudan")
```


```{r}
shapefile_df <- gadm1_map %>% 
  left_join(predictions_uncertainty, by = c("GID_1" = "gid_1")) %>% 
  left_join(predictions_wide %>% filter(year == 2025 & month == 1), by = c("GID_1" = "gid_1")) 

shapefile_df_observed <- gadm1_map %>% 
  left_join(predictions_wide %>% filter(!is.na(observed_perc_used_internet_past12months_wght_age_15_to_49_fm_ratio)), by = c("GID_1" = "gid_1"))
```



```{r}
# Women, Observed
internet_women_observed <- ggplot() +
  geom_sf(data = gadm1_map, fill = "grey70", color = "black", linewidth = 0.5) +
  geom_sf(data = shapefile_df_observed, aes(fill = observed_perc_used_internet_past12months_wght_age_15_to_49_wom), 
          color = "white", size = 0.01, alpha = 0.8, linewidth = 0.05) +
  scale_fill_viridis(na.value = "transparent", limits = c(-.001, 1), alpha = 1) +
  theme_void() +
  theme(legend.position = "bottom", plot.title = element_text(hjust = 0.3)) + 
  labs(fill = "Internet Use",
       title = "Women, observed") +
  coord_sf(
    xlim = c(-120, 160),  # Set longitude range
    ylim = c(-60, 60)     # Set latitude range
  )


# Women, error 
internet_women_predicted_error <- ggplot() +
  geom_sf(data = gadm1_map, fill = "grey70", color = "black", linewidth = 0.5) +
  geom_sf(data = shapefile_df, aes(fill = predicted_error), 
          color = "white", size = 0.01, alpha = 0.8, linewidth = 0.05) +
  scale_fill_gradientn(
    colors = viridis::magma(10), 
    values = scales::rescale(c(0, 0.1, 0.2, 0.3, 0.35)),  # Adjust values to range 0 to 0.35
    breaks = c(0, 0.1, 0.2, 0.3, 0.35),                  # Matching breaks within 0 to 0.35
    labels = c("0", "0.1", "0.2", "0.3", "0.35"),        # Labels for the breaks
    na.value = "transparent",
    guide = guide_colorbar(
      barwidth = 20,
      barheight = 1,
      title.position = "top",
      title.hjust = 0.5,
      label.hjust = 0.5
    )) + 
  theme_void() +
  theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.3, size = 14),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 10)
  ) + 
  labs(fill = "Absolute Error",
    title = "Female Internet Adoption, Absolute Error") + 
  coord_sf(
    xlim = c(-120, 160),  # Set longitude range
    ylim = c(-60, 60)     # Set latitude range
  )


# Women, relative error 
internet_women_predicted_error_relative <- ggplot() +
  geom_sf(data = gadm1_map, fill = "grey70", color = "black", linewidth = 0.5) +
  geom_sf(data = shapefile_df, aes(fill = pmin(predicted_error / predicted_perc_used_internet_past12months_wght_age_15_to_49_wom, 2)), 
          color = "white", size = 0.01, alpha = 0.8, linewidth = 0.05) +
  scale_fill_gradientn(
    colors = viridis::magma(10), 
    values = scales::rescale(c(0, 0.1, 0.3, 0.6, 2)),  # Top-coding at 2
    breaks = c(0, 0.2, 0.5, 1, 1.5, 2),                     # Adjust breaks
    labels = c("0", "0.2", "0.5", "1", "1.5", "2+"),           # Adjust labels
    na.value = "transparent",
    guide = guide_colorbar(
      barwidth = 20,
      barheight = 1,
      title.position = "top",
      title.hjust = 0.5,
      label.hjust = 0.5
    )
  ) +
  theme_void() +
  theme(
    legend.position = "bottom",
    plot.title = element_text(hjust = 0.3, size = 14),
    legend.title = element_text(size = 12, face = "bold"),
    legend.text = element_text(size = 10)
  ) + 
  labs(
    fill = "Relative Error (Error / Prediction)",
    title = "Female Internet Adoption, Relative Error"
  ) +
  coord_sf(
    xlim = c(-120, 160),  # Set longitude range
    ylim = c(-60, 60)     # Set latitude range
  )

## internet women observed and predicted 
ggsave(plot = internet_women_observed, filename = here("figures", "internet_lmic_observed.png"), height = 7, width = 10)

## create uncertainty visualization 
uncertainty_error_plot <- cowplot::plot_grid(internet_women_predicted_error, internet_women_predicted_error_relative, ncol = 1, labels = "AUTO")


ggsave(plot = uncertainty_error_plot, filename = here("figures", "internet_women_predicted_error.png"), height = 13, width = 10)
```


```{r}
titles <- c(
  "Mobile Ownership (Women, Predicted)",
  "Mobile Ownership (Men, Predicted)",
  "Mobile Ownership (Mobile Gender Gap, Predicted)",
  "Internet Use (Women, Predicted)",
  "Internet Use (Men, Predicted)",
  "Internet Use (Internet Gender Gap, Predicted)"
)

titles_legend <- c(
  "Adoption",
  "Adoption",
  "Gender Gap",
  "Adoption",
  "Adoption",
  "Gender Gap"
)

# Loop over each variable
for (i in seq_along(variables)) {
  
  var <- variables[i]
  title_text <- titles[i]
  legend_text <- titles_legend[i]
  filename <- here(paste0("figures/", var, "_map.png"))
  
  # Generate map
  plot_map <- ggplot() +
    geom_sf(data = gadm1_map, fill = "grey70", color = "black", linewidth = 0.5) +
    geom_sf(data = shapefile_df, aes_string(fill = var), 
            color = "white", size = 0.01, alpha = 0.8, linewidth = 0.05) +
    scale_fill_viridis(
      na.value = "transparent", 
      limits = c(0, 1),  # Ensure all maps use the same scale
      alpha = 1
    ) +
    theme_void() +
    theme(
      legend.position = "bottom",
      plot.title = element_text(hjust = 0.3, size = 14, face = "bold"),
      legend.margin = margin(-10, 0, 0, 0, "pt")) + 
    labs(fill = legend_text,
         title = title_text) +
    coord_sf(
      xlim = c(-120, 160),  # Set longitude range
      ylim = c(-60, 60)     # Set latitude range
    )

  # Save plot
  ggsave(plot = plot_map, filename = filename, height = 7, width = 10)
}
```



## Nigeria Plot

```{r}
shapefile_df_nigeria <- shapefile_df %>% 
  filter(COUNTRY == "Nigeria") 

gadm1_map_nigeria <- gadm1_map %>% 
  filter(COUNTRY == "Nigeria")

nigeria_predictions <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>% 
  filter(country == "Nigeria" & model == "superlearner" & feature_set == "fb_and_offline" & cv_method == "loco") %>% 
  filter(outcome == "perc_used_internet_past12months_wght_age_15_to_49_wom") %>% 
  dplyr::select(gid_1, predicted_loco = predicted, observed)


nigeria_predictions <- nigeria_predictions %>% 
  left_join(master_file_full, by = "gid_1")

shapefile_df_nigeria <- shapefile_df_nigeria %>% 
  left_join(nigeria_predictions, by = c("GID_1" = "gid_1")) %>% 
  mutate(error = observed-predicted_loco)

# Determine the common range for fill scale
fill_limits <- range(c(-.05, .55), na.rm = TRUE)

plot_truth_nigeria <- ggplot(shapefile_df_nigeria, aes(fill = observed)) + 
  geom_sf() +
  scale_fill_viridis(na.value = "transparent", limits = fill_limits, alpha = 0.8) +
  geom_sf(data = gadm1_map_nigeria, fill = "transparent", color = "black", linewidth = .5) +
  theme_void() + 
  labs(fill = "Internet, women",
       title = "Observed") 

plot_prediction_nigeria <- ggplot(shapefile_df_nigeria, aes(fill = predicted_loco)) + 
  geom_sf() +
  scale_fill_viridis(na.value = "transparent", limits = fill_limits, alpha = 0.8) +
  geom_sf(data = gadm1_map_nigeria, fill = "transparent", color = "black", linewidth = .5) +
  theme_void() + 
  labs(fill = "Internet, women",
       title = "Predicted") 

# Calculate r, R², and MAE
error_stats <- shapefile_df_nigeria %>%
  summarize(
    r = cor(observed, predicted_loco, use = "complete.obs"),
    mae = mean(abs(observed - predicted_loco)),
    .groups = "drop"
  )

# Create the plot
error_dist_nigeria <- ggplot(shapefile_df_nigeria, aes(x = observed, y = predicted_loco)) + 
  geom_point(aes(size = pop_all_age), alpha = 0.3) +
  geom_abline(slope = 1, linetype = "dashed") +
  xlim(0, 1) + 
  ylim(0, 1) + 
  theme_cowplot() + 
  labs(
    title = "",
    x = "Observed (Admin-1)",
    y = "Predicted (Admin-1)",
    size = "Population size"
  ) +
  geom_label(
    data = tibble(x = 0.2, y = 0.8, label = paste0("atop(italic(r) == ", round(error_stats$r, 2), 
                                                   ", italic(MAE) == ", round(error_stats$mae, 2), ")")),
    aes(x = x, y = y, label = label),
    size = 5,
    parse = TRUE,
    inherit.aes = FALSE
  ) +
  scale_size_continuous(
    breaks = c(2000000, 3000000, 10000000),
    labels = c("2m", "3m", "10m")
  ) +
  guides(size = guide_legend(override.aes = list(alpha = 1))) +
  theme(legend.position = "bottom")
```





```{r}
shapefile_df_zimbabwe <- shapefile_df %>% 
  filter(COUNTRY == "Zimbabwe") 

gadm1_map_zimbabwe <- gadm1_map %>% 
  filter(COUNTRY == "Zimbabwe")

zimbabwe_predictions <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>% 
  filter(country == "Zimbabwe" & model == "superlearner" & feature_set == "fb_and_offline" & cv_method == "loco") %>% 
  filter(outcome == "perc_used_internet_past12months_wght_age_15_to_49_wom") %>% 
  dplyr::select(gid_1, predicted_loco = predicted, observed)

## read in training and test datasets 
zimbabwe_predictions <- zimbabwe_predictions %>% 
  left_join(master_file_full, by = "gid_1")

shapefile_df_zimbabwe <- shapefile_df_zimbabwe %>% 
  left_join(zimbabwe_predictions, by = c("GID_1" = "gid_1")) %>% 
  mutate(error = observed-predicted_loco)

# Determine the common range for fill scale
fill_limits <- range(c(min(shapefile_df_zimbabwe$observed)-0.05, max(shapefile_df_zimbabwe$observed + 0.01)))

plot_truth_zimbabwe <- ggplot(shapefile_df_zimbabwe, aes(fill = observed)) + 
  geom_sf() +
  scale_fill_viridis(na.value = "transparent", limits = fill_limits, alpha = 0.8) +
  geom_sf(data = gadm1_map_zimbabwe, fill = "transparent", color = "black", linewidth = .5) +
  theme_void() + 
  labs(fill = "Internet, women",
       title = "Observed") 

plot_prediction_zimbabwe <- ggplot(shapefile_df_zimbabwe, aes(fill = predicted_loco)) + 
  geom_sf() +
  scale_fill_viridis(na.value = "transparent", limits = fill_limits, alpha = 0.8) +
  geom_sf(data = gadm1_map_zimbabwe, fill = "transparent", color = "black", linewidth = .5) +
  theme_void() + 
  labs(fill = "Internet, women",
       title = "Predicted") 


# Calculate r and MAE
error_stats_zim <- shapefile_df_zimbabwe %>%
  summarize(
    r = cor(observed, predicted_loco, use = "complete.obs"),
    mae = mean(abs(observed - predicted_loco)),
    .groups = "drop"
  )

# Create the plot
error_dist_zim <- shapefile_df_zimbabwe %>% 
  ggplot(aes(x = observed, y = predicted_loco)) + 
  geom_point(aes(size = pop_all_age), alpha = 0.3) +
  geom_label(
    data = tibble(
      x = 0.2, 
      y = 0.9, 
      label = paste0("atop(italic(r) == ", round(error_stats_zim$r, 2), 
                     ", italic(MAE) == ", round(error_stats_zim$mae, 2), ")")
    ),
    aes(x = x, y = y, label = label),
    size = 5.5,
    parse = TRUE,
    inherit.aes = FALSE
  ) +
  labs(
    title = "",
    x = "Observed (Admin-1)",
    y = "Predicted (Admin-1)",
    size = "Population Size"
  ) +
  theme_cowplot() + 
  xlim(0.03, 1) + 
  ylim(0.03, 1) + 
  geom_abline(slope = 1, linetype = "dashed") + 
  scale_size_continuous(
    breaks = c(1000000, 2000000),
    labels = c("1m", "2m")
  ) +
  guides(size = guide_legend(override.aes = list(alpha = 1))) +
  theme(legend.position = "bottom")

```



```{r}
shapefile_df_senegal <- shapefile_df %>% 
  filter(COUNTRY == "Senegal") 

gadm1_map_senegal <- gadm1_map %>% 
  filter(COUNTRY == "Senegal")

senegal_predictions <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>% 
  filter(country == "Senegal" & model == "superlearner" & feature_set == "fb_and_offline" & cv_method == "loco") %>% 
  filter(outcome == "perc_used_internet_past12months_wght_age_15_to_49_wom") %>% 
  dplyr::select(gid_1, predicted_loco = predicted, observed)

## read in training and test datasets 
senegal_predictions <- senegal_predictions %>% 
  left_join(master_file_full, by = "gid_1")

shapefile_df_senegal <- shapefile_df_senegal %>% 
  inner_join(senegal_predictions, by = c("GID_1" = "gid_1")) %>% 
  mutate(error = observed - predicted_loco)

# Determine the common range for fill scale
fill_limits <- range(c(min(shapefile_df_senegal$observed), max(shapefile_df_senegal$observed)))

plot_truth_senegal <- ggplot(shapefile_df_senegal, aes(fill = observed)) + 
  geom_sf() +
  scale_fill_viridis(na.value = "transparent", limits = fill_limits, alpha = 0.8) +
  geom_sf(data = shapefile_df_senegal, fill = "transparent", color = "black", linewidth = .5) +
  theme_void() + 
  labs(fill = "Internet, women",
       title = "Observed") 

plot_prediction_senegal <- ggplot(shapefile_df_senegal, aes(fill = predicted_loco)) + 
  geom_sf() +
  scale_fill_viridis(na.value = "transparent", limits = fill_limits, alpha = 0.8) +
  geom_sf(data = shapefile_df_senegal, fill = "transparent", color = "black", linewidth = .5) +
  theme_void() + 
  labs(fill = "Internet, women",
       title = "Predicted") 

# Calculate r and MAE
error_stats_senegal <- shapefile_df_senegal %>%
  summarize(
    r = cor(observed, predicted_loco, use = "complete.obs"),
    mae = mean(abs(observed - predicted_loco)),
    .groups = "drop"
  )

# Create the plot
error_dist_senegal <- shapefile_df_senegal %>% 
  ggplot(aes(x = observed, y = predicted_loco)) + 
  geom_point(aes(size = pop_all_age), alpha = 0.3) +
  geom_label(
    data = tibble(
      x = 0.2, 
      y = 0.9, 
      label = paste0("atop(italic(r) == ", round(error_stats_senegal$r, 2), 
                     ", italic(MAE) == ", round(error_stats_senegal$mae, 2), ")")
    ),
    aes(x = x, y = y, label = label),
    size = 5.5,
    parse = TRUE,
    inherit.aes = FALSE
  ) +
  labs(
    title = "",
    x = "Observed (Admin-1)",
    y = "Predicted (Admin-1)",
    size = "Population Size"
  ) +
  theme_cowplot() + 
  xlim(0.03, 1) + 
  ylim(0.03, 1) + 
  geom_abline(slope = 1, linetype = "dashed") + 
  scale_size_continuous(
    breaks = c(1000000, 2000000),
    labels = c("1m", "2m")
  ) +
  guides(size = guide_legend(override.aes = list(alpha = 1))) +
  theme(legend.position = "bottom")


```



```{r}
## read in plots create in earlier scripts 
spread_plot <- read_rds(file = here("figures", "spread_plot.rds"))
country_level_mae <- read_rds(file = here("figures", "country_level_mae.rds"))
country_level_corr <- read_rds(file = here("figures", "country_level_corr.rds"))

## create a master file 
country_error_viz <- cowplot::plot_grid(plot_truth_nigeria,
                                        plot_prediction_nigeria,
                                        error_dist_nigeria,
                                        plot_truth_zimbabwe,
                                        plot_prediction_zimbabwe,
                                        error_dist_zim, 
                                        plot_truth_senegal, 
                                        plot_prediction_senegal, 
                                        error_dist_senegal, nrow = 3, labels = "AUTO")

error_metrics <- cowplot::plot_grid(country_level_corr, country_level_mae, nrow = 2, labels = c("J", "K"))

error_plot <- cowplot::plot_grid(error_metrics, spread_plot, labels = c("", "L")) 

country_error_viz_final <- plot_grid(country_error_viz, error_plot, ncol = 1, rel_heights = c(1, 0.5))

ggsave(plot = country_error_viz_final, filename = here("figures", "country_error_viz.png"), height = 17, width = 14)

ggsave(plot = country_error_viz_final, filename = here("figures", "fig3_country_error_viz.pdf"), height = 17, width = 14)

```
