==> code/06_feature_importance.Rmd <==
---
title: "04 - Quantify R2"
auhor: "Casey Breen" 
---

Summary: Generate figures and tables summarizing most important features 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```

## Estimate model performance metrics 

```{r}
## read predictions
predictions_combined <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>% 
  mutate(predicted = case_when(predicted < 0 ~ 0,
                  TRUE ~ predicted)) %>% 
  mutate(cv_method = case_when(
    grepl("10fold", cv_method) ~ "10-Fold-CV",
    grepl("loco", cv_method) ~ "LOCO-CV",
    TRUE ~ cv_method
  )) %>% 
  mutate(outcome_categorical = case_when(
    outcome %in% c("perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio",
                   "perc_owns_mobile_telephone_wght_age_15_to_49_men",
                   "perc_owns_mobile_telephone_wght_age_15_to_49_wom") ~ "Mobile",
    outcome %in% c("perc_used_internet_past12months_wght_age_15_to_49_fm_ratio",
                   "perc_used_internet_past12months_wght_age_15_to_49_men",
                   "perc_used_internet_past12months_wght_age_15_to_49_wom") ~ "Internet"))

## read in training and test datasets 
master_file <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>%
  dplyr::select(country, gid_1, dhsyear, pop_all_age_total, continent) 


predictions_combined <- predictions_combined %>% 
  left_join(master_file, by = "gid_1")

## estimate model performance metrics 
## r2, rmse, mae are fine 
## mape and mare are a bit tricky because observed value is sometimes 0 
model_performance_metrics <- predictions_combined %>%
  group_by(cv_method, model, outcome, feature_set) %>%
  summarise(r2 = compute_r2(predicted, observed),
            rmse = compute_rmse(predicted, observed),
            mape = compute_mape(predicted, observed),
            mare = compute_mare(predicted, observed),
            mae = compute_mae(predicted, observed)) %>%
  pivot_longer(cols = c(r2, rmse, mape, mare, mae), 
               names_to = "metric", 
               values_to = "value")

## model metrics 
model_performance_metrics <- model_performance_metrics %>% 
  mutate(outcome_clean = case_when(
    grepl("mobile_telephone_wght_age_15_to_49_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_telephone_wght_age_15_to_49_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_telephone_wght_age_15_to_49_wom", outcome) ~ "Mobile (Women)",
    grepl("used_internet_past12months_wght_age_15_to_49_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("used_internet_past12months_wght_age_15_to_49_men", outcome) ~ "Internet (Men)",
    grepl("used_internet_past12months_wght_age_15_to_49_wom", outcome) ~ "Internet (Women)",
    TRUE ~ outcome))  # Default to keep original name if no match
```

## R-Squared Plot

```{r}
# Define the desired order of facets
desired_order <- c("Internet (Men)", "Internet (Women)", "Internet (Ratio)",
                   "Mobile (Men)", "Mobile (Women)", "Mobile (Ratio)")

## r2 plot -- algo X outcome
r2_plot <- model_performance_metrics %>% 
  filter(metric == "r2" & feature_set == "fb_and_offline") %>% 
  filter(!outcome %in% c("perc_ever_used_internet_wght_age_15_to_49_fm_ratio",
                         "perc_ever_used_internet_wght_age_15_to_49_men",
                         "perc_ever_used_internet_wght_age_15_to_49_wom")) %>% 
  mutate(value = round(value, 2)) %>% 
  mutate(outcome_clean = factor(outcome_clean, levels = desired_order)) %>%  # Set the factor levels
  ggplot(aes(x = reorder(model, -value), y = value, fill = cv_method, label = value)) + 
  geom_col(position = "dodge", color = "black") + 
  geom_text(position = position_dodge(width = .9), vjust = -0.5, size = 4) + 
  scale_y_continuous(limits = c(0, 1)) + 
  theme_cowplot() + 
  labs(
    x = "Algorithm",
    y = expression(bold(R^2)),
    title = expression(bold("Algorithm Performance " (R^2))),
    fill = "Cross-Validation") + 
  facet_wrap(~outcome_clean, ncol = 1) + 
  scale_fill_manual(values=cbp1) + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(1.5, "lines"),
    legend.position = "bottom",
    legend.justification = "center",  # centers the whole legend box
    legend.title = element_text(size = 13, hjust = 0.5, vjust = 1, margin = margin(b = 4)),
    legend.text = element_text(size = 11)
  ) 

## save plot 
ggsave(plot = r2_plot, filename = here("figures", "r2_plot_dhsyear.png"), height = 12, width = 8)
```



## feature comparison 

```{r}
## facet plots
filtered_data <- model_performance_metrics %>%
  filter(metric == "r2" & cv_method == "LOCO-CV" & model == "superlearner") %>%
  mutate(
    category = case_when(
      grepl("owns_mobile", outcome) ~ "Own Mobible",
      grepl("used_internet_past12months", outcome) ~ "Used Internet (12m)",
      grepl("ever_used_internet", outcome) ~ "Used Internet (ever)",
      TRUE ~ "Other"
    ),
    gender_group = case_when(
      grepl("_men$", outcome) ~ "Men",
      grepl("_wom$", outcome) ~ "Women",
      grepl("_fm_ratio$", outcome) ~ "Gender Gap Index",
      TRUE ~ "Other"
    )
  )

feature_comparison_plot <- filtered_data %>%
  mutate(feature_set = case_when(
    feature_set == "fb" ~ "Facebook",
    feature_set == "fb_and_offline" ~ "Facebook and Offline",
    feature_set == "offline" ~ "Offline"
  )) %>% 
  ggplot() + 
  geom_point(aes(x = value, y = gender_group, color = feature_set, shape = feature_set), size = 5, position = position_dodge(width = .25)) +
  theme_minimal(16) +
  labs(x = expression("Model Performance: R" ^ 2),
       y = NULL
  ) +
  theme(
    legend.position = c(0.1, 0.97),
    legend.justification = c(0, 1),
    legend.box.background = element_rect(color = "black", fill = "white", linewidth = 0.3),
    legend.background = element_blank(),
    panel.border = element_rect(colour = "black", fill = NA, size = 1),
    panel.grid = element_line(color = "grey95"),
    strip.background = element_rect(colour = "black", fill = "white",  linewidth = 1),
    strip.text.y = element_text(angle = 0, color = "black", hjust = 0.5),
    strip.placement = "inside",
    legend.key.size = unit(0.3, "cm"),  # Adjusts size of legend keys
    legend.text = element_text(size = 12),  # Adjusts size of legend text
    legend.title = element_blank()) +
 # ggsci::scale_color_lancet() + 
  scale_color_viridis_d() + 
  xlim(0, 0.8) + 
  facet_grid(category ~ ., scales = "free_y", space = "free_y", switch = "y")
```



```{r}
## read in training and test datasets 
master_file <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>%
  dplyr::select(country, gid_1, all_of(c(dhs_vars, fb_feats, off_feats))) %>% ## the vectors of vars are defined in helper function
  filter(!is.na(perc_used_internet_past12months_wght_age_15_to_49_wom)) %>% drop_na() %>% 
  mutate(continent = case_when(continent == "Africa" ~ 1,
                               TRUE ~ 0)) 

# Define your features and dependent variable
features <- tolower(c(fb_feats, off_feats)) 
dependent_vars <- c(
  "perc_used_internet_past12months_wght_age_15_to_49_wom",
  "perc_owns_mobile_telephone_wght_age_15_to_49_wom",
  "perc_owns_mobile_telephone_wght_age_15_to_49_men",
  "perc_used_internet_past12months_wght_age_15_to_49_men",
  "perc_used_internet_past12months_wght_age_15_to_49_fm_ratio",
  "perc_owns_mobile_telephone_wght_age_15_to_49_fm_ratio"
)
country_col <- "country" # Adjust this to match the country identifier in your dataset

# Loop through each dependent variable
cv_results_list <- list()
for (dependent_var in dependent_vars) {
  # Ensure the data subset does not include missing values for the current dependent variable
  data_subset <- filter(master_file, !is.na(!!sym(dependent_var)))

  # Perform cross-validation
  cv_results <- perform_cross_validation_loco(data_subset, dependent_var, country_col = country_col, features) #

  # Store the results
  cv_results_list[[dependent_var]] <- cv_results %>% 
    mutate(outcome = dependent_var)
}

# Print results for each dependent variable
cv_results <- bind_rows(cv_results_list)
```




```{r}
# Assuming cv_results is your dataset
var_importance_data <- cv_results %>%
  group_by(outcome) %>%
  arrange(desc(avg_r_squared)) %>%
  slice_head(n = 10) %>% 
  mutate(feature_recode = case_when(
    feature == "fb_pntr_18p_female" ~ "Subnational Facebook Penetration (Female)",
    feature == "fb_pntr_18p_male" ~ "Subnational Facebook Penetration (Male)",
    grepl("all_devices_age_18_plus_gg", feature) ~ "Subnational Facebook Penetration Gender Gap",
    grepl("fb_pntr_18p_male_national", feature) ~ "National Facebook Penetration (Male)",
    grepl("fb_pntr_18p_female_national", feature) ~ "National Facebook Penetration (Female)",
    grepl("rwi_mean_wght", feature) ~ "Subnational Relative Wealth Index (Pop-weighted)",
    grepl("subnational_hdi_males", feature) ~ "Subnational Human Development Index (Male)",
    grepl("rwi_mean", feature) ~ "Subnational Relative Wealth Index (Mean)",
    grepl("subnational_hdi_females", feature) ~ "Subnational Human Development Index (Female)",
    grepl("ios_age_18_plus_female_frac_2024", feature) ~ "Subnational iOS FB Usage (Female)",
    grepl("income_index_males", feature) ~ "Subnational Income Index (Male)",
    grepl("educational_index_females", feature) ~ "Subnational Educational Index (Female)",
    grepl("educational_index_males", feature) ~ "Subnational Educational Index (Male)",
    grepl("nl_mean", feature) ~ "Subnational Night Light Mean",
    grepl("wifi_age_18_plus_female_frac_2024_tess", feature) ~ "Subnational WiFi FB Usage (Female)",
    grepl("wifi_age_18_plus_male_frac_2024_tess", feature) ~ "Subnational WiFi FB Usage (Male)",
    grepl("×4g_age_13_plus_female_frac_2024", feature) ~ "Subnational WiFi FB Usage (Male)",
    grepl("hdi", feature) ~ "National Human Development Index",
    feature == "gdi" ~ "National Gender Development Index",
    feature == "subnational_gdi" ~ "Subnational Gender Development Index",
    grepl("ios_age_18_plus_male_frac_2024", feature) ~ "Subnational iOS FB Usage (Male)",
    grepl("income_index_females", feature) ~ "Subnational Income Index (Female)",
    grepl("pop_density", feature) ~ "Subnational Population Density",
    grepl("wifi_age_18_plus_male_frac_2021", feature) ~ "Subnational WiFi FB Usage (Male)",
    grepl("wifi_age_18_plus_female_frac_2021", feature) ~ "Subnational WiFi FB Usage (Female)",
    feature == "ios_age_13_plus_male_frac_2024" ~ "Subnational iOS FB Usage (Male)",
    feature == "dhsyear" ~ "DHS Survey Year",
    feature == "x4g_age_13_plus_female_frac_2024" ~ "Subnational 4G+ Network FB Usage", 
    grepl("all_devices_age_18_dhsyear_gg_national", feature) ~ "Subnational Facebook National Gender Gap",
    grepl("all_devices_age_18_plus_gg_2021_impute", feature) ~ "Subnational Facebook Gender Gap",
    grepl("rel_date", feature) ~ "Year",
    TRUE ~ feature  # Default case
  )) %>% 
  group_by(outcome) %>%
  arrange(avg_r_squared) %>%
  mutate(feature_idea = paste0(outcome, '_', feature_recode)) %>% 
  mutate(feature_idea = factor(feature_idea, levels = unique(feature_idea))) %>%
  ungroup() %>% 
  mutate(outcome_clean = case_when(
    grepl("internet_wght_age_15_to_49_fm_ratio", outcome) ~ "Internet, Ever (Ratio)",
    grepl("internet_wght_age_15_to_49_men", outcome) ~ "Internet, Ever (Men)",
    grepl("internet_wght_age_15_to_49_wom", outcome) ~ "Internet, Ever (Women)",
    grepl("mobile_telephone_wght_age_15_to_49_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_telephone_wght_age_15_to_49_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_telephone_wght_age_15_to_49_wom", outcome) ~ "Mobile (Women)",
    grepl("used_internet_past12months_wght_age_15_to_49_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("used_internet_past12months_wght_age_15_to_49_men", outcome) ~ "Internet (Men)",
    grepl("used_internet_past12months_wght_age_15_to_49_wom", outcome) ~ "Internet (Women)",
    TRUE ~ outcome
  )) # Default to keep original name if no match


var_importance_plot <- var_importance_data %>% 
  filter(outcome == "perc_used_internet_past12months_wght_age_15_to_49_wom") %>% 
  ggplot(aes(y = feature_idea, x = avg_r_squared)) +
  geom_point(size = 3) +
  theme_cowplot() +
  geom_segment(aes(yend = feature_idea, x = 0, xend = avg_r_squared), linewidth = 0.5) +
 # geom_segment(aes(yend = feature_idea, x = 0, xend = max(avg_r_squared)), linetype = "dashed", size = 0.1) +
  labs(x = expression("R" ^ 2),
       y = "",
       title = "Feature Importance") +
  # facet_wrap(~outcome_clean, scales = "free_y") +
  scale_y_discrete(breaks = var_importance_data$feature_idea,
                   labels = var_importance_data$feature_recode)
```



```{r}
r_squaredplot <- plot_grid(feature_comparison_plot, var_importance_plot, labels = "AUTO")

ggsave(plot = r_squaredplot, filename = here("figures", "feature_and_var_importance.png"), width = 13, height = 6, units = "in", dpi = 300)

ggsave(plot = r_squaredplot, filename = here("figures", "fig6_feature_and_var_importance.pdf"), width = 13, height = 6, dpi = 300)
```



```{r}
var_importance_data$outcome_clean_f = factor(var_importance_data$outcome_clean, 
                                             levels=c('Internet (Women)',
                                                      'Internet (Men)',
                                                      'Mobile (Women)',
                                                      'Mobile (Men)',
                                                      'Internet (Ratio)',
                                                      'Mobile (Ratio)'))

var_importance_plot <- var_importance_data %>% 
  ggplot(aes(y = feature_idea, x = avg_r_squared)) +
  geom_point() +
  theme_cowplot() +
  geom_segment(aes(yend = feature_idea, x = 0, xend = avg_r_squared), size = 0.5) +
 # geom_segment(aes(yend = feature_idea, x = 0, xend = max(avg_r_squared)), linetype = "dashed", size = 0.1) +
  labs(x = expression("R" ^ 2),
       y = "",
       title = "Feature Importance") +
  facet_wrap(~outcome_clean_f, scales = "free_y", nrow = 3) +
  scale_y_discrete(breaks = var_importance_data$feature_idea,
                   labels = var_importance_data$feature_recode) + 
   theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(1.5, "lines"),
    aspect.ratio = .9,
    legend.position = "bottom",
    legend.justification = "center",  # centers the whole legend box
    legend.title = element_text(size = 13, hjust = 0.5, vjust = 1, margin = margin(b = 4)),
    legend.text = element_text(size = 11)
  ) 

ggsave(plot = var_importance_plot, filename = here( "figures", "feature_importance.png"), width = 13, height = 8, units = "in", dpi = 300)
```
