==> code/08_residual_analysis.Rmd <==
---
title: "Residual Analysis"
author: Casey Breen
---

Summary: Calculate and plot model residuals 

## Library packages 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```

## Read in data files 

```{r}
## read in predictions 
predictions_combined <- read_csv(here("out", "combined_predictions_lmics_national_covars_full.csv")) %>% 
  mutate(predicted = case_when(predicted < 0 ~ 0,
                  TRUE ~ predicted)) %>% 
  filter(cv_method == "loco") %>% 
  filter(feature_set == "fb_and_offline") 

## read in training and test datasets 
master_file <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>% 
  dplyr::select(country, gid_1, all_of(c(dhs_vars, fb_feats, off_feats))) %>% ## the vectors of vars are defined in helper function 
  filter(!is.na(perc_used_internet_past12months_wght_age_15_to_49_wom)) %>%
  drop_na()
```


```{r}
predictions_filter <- predictions_combined %>% 
  dplyr::select(-country) %>% 
  filter(model == "superlearner") 
```



```{r}
predictions_filter <- predictions_filter %>% 
  left_join(master_file, by = "gid_1")

predictions_filter <- predictions_filter %>% 
  mutate(residual = observed - predicted) %>% 
  mutate(mean_residual = mean(residual))
```


```{r}
## predictions filter
predictions_filter <- predictions_filter %>% mutate(outcome_clean = factor(outcome_clean, 
                                levels = c("Internet (Women)", "Internet (Men)", "Internet (Ratio)",
                                           "Mobile (Women)", "Mobile (Men)", "Mobile (Ratio)")))

# Combined plot using facet_grid
combined_plot_residuals <- predictions_filter %>%
  ggplot() +
  geom_histogram(aes(x = residual), bins = 50, fill = "grey", color = "black") +
  geom_vline(aes(xintercept = mean_residual), linetype = "dashed", color = "red", linewidth = 1) +
  facet_wrap(~outcome_clean) +
  theme_cowplot() +
  labs(x = "Residual", y = "Count") +
  xlim(-1, 1) + 
  theme(
      strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
      strip.text = element_text(size = 9, face = "bold", color = "black"),
      panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
      panel.background = element_rect(fill = "#f9f9f9", color = NA),
      panel.grid = element_line(color = "grey95"),
      panel.spacing = unit(.2, "lines"),
      aspect.ratio = .9,
      axis.text = element_text(size = 11),            # Axis tick labels
      legend.position = "bottom")

ggsave(plot = combined_plot_residuals, filename = here("figures", "combined_plot_residuals.png"), height = 5, width = 8)
```
