==> code/09_var_explained.Rmd <==
---
title: "Within vs. between country"
author: Casey Breen
---

Summary: Analysis of within vs. between country variance 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```


```{r}
## read in training and test datasets
master_file <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>%
  filter(year == 2025 & month == 01) %>% 
  dplyr::select(country, all_of(c(dhs_vars, fb_feats, off_feats))) %>%
  filter(!is.na(perc_used_internet_past12months_wght_age_15_to_49_wom)) %>% drop_na() 

master_file_full <- read_csv(here("data", "master_data_file_march12_with_national.csv")) %>%
  filter(year == 2025 & month == 01) %>% 
   filter(!is.na(fb_pntr_18p_female)) 
```


```{r}
predictions <- read_csv(here("out", "predictions_wide.csv")) %>% 
  filter(year == 2025 & month == 01)

predictions <- predictions %>% 
  left_join(master_file_full %>% select(-country), by = c("gid_1" = "gid_1")) %>% 
  mutate(hdi_bin = ntile(hdi, 4)) 
```

## save file 

```{r}
# Function to calculate MLD and its components
calculate_mld_components <- function(data, rate_col, constant = 0.5) {
  
  # Convert rate_col and country_col to symbols for NSE
  rate_col_sym <- sym(rate_col)

  # Add a small constant to avoid logarithm of zero
  data <- data %>%
    mutate(!!rate_col_sym := !!rate_col_sym + constant)
  
  # Calculate overall mean rate (mu)
  overall_mean <- mean(data[[rate_col]], na.rm = TRUE)
  
  # Function to calculate MLD for a given vector of rates
  calc_mld <- function(rates, mean_rate) {
    n <- length(rates)
    mld <- sum(log(mean_rate / rates), na.rm = TRUE) / n
    return(mld)
  }
  
  # Calculate within-country MLD
  within_mlds <- data %>%
    group_by(country) %>%
    summarise(
      N_k = n(),
      mu_k = mean(!!rate_col_sym, na.rm = TRUE),
      MLD_k = calc_mld(!!rate_col_sym, mu_k),
      .groups = 'drop'
    )
  
  # Calculate within-country MLD component (MLD_W)
  within_mlds <- within_mlds %>%
    mutate(weight = N_k / sum(N_k, na.rm = TRUE)) %>%
    mutate(weighted_MLD_k = weight * MLD_k)
  
  MLD_W <- sum(within_mlds$weighted_MLD_k, na.rm = TRUE)
  
  # Calculate between-country MLD component (MLD_B)
  between_mlds <- within_mlds %>%
    mutate(MLD_B_component = weight * log(overall_mean / mu_k))
  
  MLD_B <- sum(between_mlds$MLD_B_component, na.rm = TRUE)
  
  # Calculate total MLD
  total_MLD <- MLD_W + MLD_B
  
  # Print results
  cat("Within-country MLD (MLD_W):", MLD_W, "\n")
  cat("Between-country MLD (MLD_B):", MLD_B, "\n")
  cat("Total MLD:", total_MLD, "\n")
  cat("Ratio of Within-country to Between-country MLD:", MLD_W / MLD_B, "\n")
  
  # Return results as a list
  return(data.frame(MLD_W = MLD_W, 
                    MLD_B = MLD_B,
                    total_MLD = total_MLD,
                    proportion_within = MLD_W / total_MLD ))
}


hdi_bins <- c(1, 2, 3, 4)
result_list <- list()

# List of columns to calculate MLD for
predicted_cols <- c(
  "predicted_perc_owns_mobile_telephone_wght_age_15_to_49_wom",
  "predicted_perc_owns_mobile_telephone_wght_age_15_to_49_men",
  "predicted_perc_used_internet_past12months_wght_age_15_to_49_wom",
  "predicted_perc_used_internet_past12months_wght_age_15_to_49_men"
)

## store results 
for (hdi_bin in hdi_bins) {
  for (col in predicted_cols) {
    
    result <- calculate_mld_components(predictions %>% filter(hdi_bin == !!hdi_bin), 
                                       rate_col = col) %>% 
      mutate(hdi_bin = hdi_bin) %>% 
      mutate(col = col)
    
    key <- paste0(hdi_bin, "_", col)
    
    result_list[[key]] <- result
  
} 
}

result_list_all <- list()

for (col in predicted_cols) {
  overall_result <- calculate_mld_components(predictions, 
                                             rate_col = col) %>%
    mutate(hdi_bin = "Overall", col = col)
  
  key <- paste0("Overall_", col)
  result_list_all[[key]] <- overall_result
}

result_list_df <- bind_rows(result_list) %>% 
  mutate(hdi_bin = case_when(
    hdi_bin == 1 ~ "HDI Q1 (Lowest)",
    hdi_bin == 2 ~ "HDI Q2",
    hdi_bin == 3 ~ "HDI Q3",
    hdi_bin == 4 ~ "HDI Q4 (Highest)"))


result_list_all_df <- bind_rows(result_list_all)

results_full <- bind_rows(result_list_df, result_list_all_df)


results_full <- results_full %>% 
  mutate(outcome_clean = case_when(
    grepl("mobile_telephone_wght_age_15_to_49_fm_ratio", col) ~ "Mobile (Ratio)",
    grepl("mobile_telephone_wght_age_15_to_49_men", col) ~ "Mobile (Men)",
    grepl("mobile_telephone_wght_age_15_to_49_wom", col) ~ "Mobile (Women)",
    grepl("used_internet_past12months_wght_age_15_to_49_fm_ratio", col) ~ "Internet (Ratio)",
    grepl("used_internet_past12months_wght_age_15_to_49_men", col) ~ "Internet (Men)",
    grepl("used_internet_past12months_wght_age_15_to_49_wom", col) ~ "Internet (Women)",
    TRUE ~ col
  )) 
```



```{r}
results_full$cat <- factor(results_full$hdi_bin, levels = c("HDI Q1 (Lowest)", "HDI Q2", "HDI Q3", "HDI Q4 (Highest)", "Overall"))

var_explained_fig <- results_full %>%
  ggplot(aes(x = outcome_clean, y = proportion_within, fill = as.factor(cat))) +
  geom_bar(stat = "identity", width = .75, position = position_dodge(width = .75), color = "black") +
  geom_text(aes(label = round(proportion_within* 100, 0)), 
            position = position_dodge(width = .75), vjust = -0.5, hjust = .5, size = 3.5) +
  scale_fill_manual(values = cud) +
  theme_cowplot() +
#scale_y_continuous(labels = scales::percent_format(accuracy = 5L), limits = c(0, .55)) + 
  theme(legend.position = "bottom",
        legend.text = element_text(size = 10)) +
  labs(fill = "",
       x = "",
       y = "Within-country contribution \n (% total variation)") +
     background_grid() + 
  geom_hline(yintercept = 0.5, linetype = "dashed", color = "black") + 
  ylim(0, .65)

ggsave(var_explained_fig, filename = here("figures", "var_explained_fig.png"), height = 4, width = 7)

ggsave(var_explained_fig, filename = here("figures", "fig4_var_explained.pdf"), height = 4, width = 7)
```
