==> code/11_disparities_plots.Rmd <==
---
title: "Disparities plots"
author: Casey Breen
---

Summary: Make a plot about disparities

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```


```{r}
predictions <- read_csv(here("out", "predictions_with_uncertainty.csv"))

predictions <- predictions %>% 
  mutate(date = as.Date(paste0(date, "-01")))
```



```{r}
abs_diff_results <- predictions %>% 
  filter(date == "2025-01-01" & outcome %in% c("internet_women", "internet_men", 'mobile_men', "mobile_women")) %>% 
  group_by(gid_0, outcome) %>% 
  mutate(predicted = case_when(
    predicted > 1 ~ 1,
    TRUE ~ predicted)) %>% 
  summarize(diff = max(predicted) - min(predicted),
            avg = median(predicted),
            min = min(predicted),
            max = max(predicted)) 

abs_diff_results_mean <-  abs_diff_results %>% 
  group_by(outcome) %>% 
  summarize(diff = mean(max) - mean(min),
            min = mean(min),
            max = mean(max))  %>% 
  mutate(country = "Overall mean")  %>%
  mutate(
    gender = ifelse(str_detect(outcome, "women"), "Women", "Men")) 
        
```


```{r}
# Prepare Internet data for dumbbell plot
internet_data <- abs_diff_results %>%
  filter(str_detect(outcome, "internet")) %>%
  mutate(
    gender = ifelse(str_detect(outcome, "women"), "Women", "Men"),
    country = countrycode(gid_0, origin = "iso3c", destination = "country.name")
  ) %>%
  mutate(country = case_when(
    country == "Congo - Brazzaville" ~ "Republic of the Congo", 
    country == "Congo - Kinshasa" ~ "Democratic Republic of the Congo",
    TRUE ~ country
  )) %>% 
  group_by(outcome) %>%
  slice_max(diff, n = 10, with_ties = FALSE)  %>% # Select top 10 countries by gender
  bind_rows(abs_diff_results_mean) %>% 
  filter(str_detect(outcome, "internet")) 

# Prepare Mobile data for dumbbell plot
mobile_data <- abs_diff_results %>%
  filter(str_detect(outcome, "mobile")) %>%
  mutate(
    gender = ifelse(str_detect(outcome, "women"), "Women", "Men"),
    country = countrycode(gid_0, origin = "iso3c", destination = "country.name")
  ) %>%
  mutate(country = case_when(
    country == "Congo - Brazzaville" ~ "Republic of the Congo", 
    country == "Congo - Kinshasa" ~ "Democratic Republic of the Congo",
    TRUE ~ country
  )) %>% 
  group_by(outcome) %>%
  slice_max(diff, n = 10, with_ties = FALSE) %>%  # Select top 10 countries by gender
  bind_rows(abs_diff_results_mean) %>% 
  filter(str_detect(outcome, "mobile")) 
```




```{r}
## make data frames 
mobile_data_men <- mobile_data %>% filter(gender == "Men")
mobile_data_women <- mobile_data %>% filter(gender == "Women")

# Men plot
mobile_plot_men <- mobile_data_men %>%
  mutate(diff = max - min) %>%
  mutate(
    country = fct_relevel(country, "Overall mean"), # Move "Overall mean" to the top
    country = fct_reorder(country, diff, .desc = FALSE) # Order by descending difference
  ) %>%
  ggplot(aes(y = country, x = max, xend = min)) +
  geom_dumbbell(
    color = "darkgray",
    size = 1,
    size_x = 3,
    size_xend = 3,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  geom_dumbbell(
    data = mobile_data_men %>% filter(country == "Overall mean"),
    aes(y = fct_relevel(country, "Overall mean"), x = max, xend = min),
    color = "darkgrey",
    size = 1.5,
    size_x = 4,
    size_xend = 4,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  labs(
    x = "",
    y = "",
    title = "Men"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.y = element_text(size = 10, face = "bold"),
    axis.title.x = element_text(size = 14, face = "bold")
  ) + 
  scale_y_discrete(
    labels = c("Overall mean" = expression(bold("Overall mean")), "Other Country" = "Other Country"),
  ) + 
  # Add numbers above each bar
geom_text(aes(x = (max + min) / 2-0.05, y = country, label = round(max-min, 2)), 
          vjust = -0.5, hjust = 0, size = 3.5, color = "black") +
  xlim(0, 1)

# Women plot
mobile_plot_women <- mobile_data_women %>%
  mutate(diff = max - min) %>%
  mutate(
    country = fct_relevel(country, "Overall mean"), # Move "Overall mean" to the top
    country = fct_reorder(country, diff, .desc = FALSE) # Order by descending difference
  ) %>%
  ggplot(aes(y = country, x = max, xend = min)) +
  geom_dumbbell(
    color = "darkgray",
    size = 1,
    size_x = 3,
    size_xend = 3,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  geom_dumbbell(
    data = mobile_data_women %>% filter(country == "Overall mean"),
    aes(y = fct_relevel(country, "Overall mean"), x = max, xend = min),
    color = "darkgrey",
    size = 1.5,
    size_x = 4,
    size_xend = 4,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  labs(
    x = "Mobile Adoption",
    y = "",
    title = "Women"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.y = element_text(size = 10, face = "bold"),
    axis.title.x = element_text(size = 14, face = "bold")
  ) + 
  # Use ggtext to bold "Overall mean" via Markdown
  scale_y_discrete(
    labels = c("Overall mean" = expression(bold("Overall mean")), "Other Country" = "Other Country")
  ) + 
  xlim(0, 1) + 
  # Add numbers above each bar
geom_text(aes(x = (max + min) / 2-0.05, y = country, label = round(max-min, 2)), 
          vjust = -0.5, hjust = 0, size = 3.5, color = "black") +
  xlim(0, 1)
```

```{r}
## make data frames 
internet_data_men <- internet_data %>% filter(gender == "Men")
internet_data_women <- internet_data %>% filter(gender == "Women")

# Men plot
internet_plot_men <- internet_data_men %>%
  mutate(diff = max - min) %>%
  mutate(
    country = fct_relevel(country, "Overall mean"), # Move "Overall mean" to the top
    country = fct_reorder(country, diff, .desc = FALSE) # Order by descending difference
  ) %>%
  ggplot(aes(y = country, x = max, xend = min)) +
  geom_dumbbell(
    color = "darkgray",
    size = 1,
    size_x = 3,
    size_xend = 3,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  geom_dumbbell(
    data = internet_data_men %>% filter(country == "Overall mean"),
    aes(y = fct_relevel(country, "Overall mean"), x = max, xend = min),
    color = "darkgrey",
    size = 1.5,
    size_x = 4,
    size_xend = 4,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  labs(
    x = "",
    y = "",
    title = "Men"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.y = element_text(size = 10, face = "bold"),
    axis.title.x = element_text(size = 14, face = "bold")
  ) + 
  # Use ggtext to bold "Overall mean" via Markdown
  scale_y_discrete(
    labels = c("Overall mean" = expression(bold("Overall mean")), "Other Country" = "Other Country")
  ) + 
  xlim(0, 1)+ 
  # Add numbers above each bar
geom_text(aes(x = (max + min) / 2-0.05, y = country, label = round(max-min, 2)), 
          vjust = -0.5, hjust = 0, size = 3.5, color = "black") +
  xlim(0, 1)

# Women plot
internet_plot_women <- internet_data_women %>%
  mutate(diff = max - min) %>%
  mutate(
    country = fct_relevel(country, "Overall mean"), # Move "Overall mean" to the top
    country = fct_reorder(country, diff, .desc = FALSE) # Order by descending difference
  ) %>%
  ggplot(aes(y = country, x = max, xend = min)) +
  geom_dumbbell(
    color = "darkgray",
    size = 1,
    size_x = 3,
    size_xend = 3,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  geom_dumbbell(
    data = internet_data_women %>% filter(country == "Overall mean"),
    aes(y = fct_relevel(country, "Overall mean"), x = max, xend = min),
    color = "darkgrey",
    size = 1.5,
    size_x = 4,
    size_xend = 4,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  labs(
    x = "Internet Adoption ",
    y = "",
    title = "Women"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.y = element_text(size = 10, face = "bold"),
    axis.title.x = element_text(size = 14, face = "bold")
  ) + 
  # Use ggtext to bold "Overall mean" via Markdown
  scale_y_discrete(
    labels = c("Overall mean" = expression(bold("Overall mean")), "Other Country" = "Other Country")
  ) + 
  xlim(0, 1) + 
  # Add numbers above each bar
geom_text(aes(x = (max + min) / 2-0.05, y = country, label = round(max-min, 2)), 
          vjust = -0.5, hjust = 0, size = 3.5, color = "black") +
  xlim(0, 1)

```



```{r}
# Combine plots using cowplot
combined_plot <- plot_grid(mobile_plot_men, internet_plot_men,  mobile_plot_women, internet_plot_women, ncol = 2, labels = "AUTO")

# Dummy plot for shared legend
legend_plot_levels <- ggplot(
  data.frame(x = c(1, 2), y = c(1, 2), group = c("Subnational Region with Lowest Adoption", "Subnational Region with Highest Adoption")),
  aes(x = x, y = y, color = group)
) +
  geom_point(size = 4) +
  scale_color_manual(
    values = c(
      "Subnational Region with Lowest Adoption" = "#699DC6",
      "Subnational Region with Highest Adoption" = "#F69541"
    ),
    breaks = c("Subnational Region with Lowest Adoption", "Subnational Region with Highest Adoption")
  ) +
  guides(color = guide_legend(title = NULL)) +
  theme_void() +
  theme(legend.position = "bottom", legend.text = element_text(size = 10))

## Extract Legend 
legend_levels <- ggpubr::get_legend(legend_plot_levels)

## extract legend 
combined_plot_levels_legend <- plot_grid(combined_plot, legend_levels, ncol = 1, rel_heights = c(2, 0.1))

## ggsave 
ggsave(plot = combined_plot_levels_legend, filename = here("figures", "withincountry_disparities_levels.png"), width = 10, height = 8)
```




```{r}
abs_diff_results_gaps <- predictions %>% 
  filter(date == "2025-01-01" & outcome %in% c("internet_fm_ratio", "mobile_fm_ratio")) %>% 
  group_by(gid_0, outcome) %>% 
  mutate(predicted = case_when(
    predicted > 1 ~ 1,
    TRUE ~ predicted)) %>% 
  summarize(diff = max(predicted) - min(predicted),
            avg = median(predicted),
            min = min(predicted),
            max = max(predicted)) 

abs_diff_results_gaps_mean <-  abs_diff_results_gaps %>% 
  group_by(outcome) %>% 
  summarize(diff = mean(max) - mean(min),
            min = mean(min),
            max = mean(max))  %>% 
  mutate(country = "Overall mean")
```

```{r}
# Prepare Internet data for dumbbell plot
gender_gap_data_internet <- abs_diff_results_gaps %>%
  mutate(
    country = countrycode(gid_0, origin = "iso3c", destination = "country.name")
  ) %>%
  mutate(country = case_when(
    country == "Congo - Brazzaville" ~ "Republic of the Congo", 
    country == "Congo - Kinshasa" ~ "Democratic Republic of the Congo",
    TRUE ~ country
  )) %>% 
  group_by(outcome) %>%
  slice_max(diff, n = 10, with_ties = FALSE)  %>% # Select top 10 countries by gender
  bind_rows(abs_diff_results_gaps_mean) %>% 
  filter(str_detect(outcome, "internet")) 


# Prepare Internet data for dumbbell plot
gender_gap_data_mobile <- abs_diff_results_gaps %>%
  mutate(
    country = countrycode(gid_0, origin = "iso3c", destination = "country.name")
  ) %>%
  mutate(country = case_when(
    country == "Congo - Brazzaville" ~ "Republic of the Congo", 
    country == "Congo - Kinshasa" ~ "Democratic Republic of the Congo",
    TRUE ~ country
  )) %>% 
  group_by(outcome) %>%
  slice_max(diff, n = 10, with_ties = FALSE)  %>% # Select top 10 countries by gender
  bind_rows(abs_diff_results_gaps_mean) %>% 
  filter(str_detect(outcome, "mobile"))
```


```{r}
## Gender Gap Plot, internet 

gender_gap_internet <- gender_gap_data_internet %>%
  mutate(diff = max - min) %>%
  mutate(
    country = fct_relevel(country, "Overall mean"), # Move "Overall mean" to the top
    country = fct_reorder(country, diff, .desc = FALSE) # Order by descending difference
  ) %>%
  ggplot(aes(y = country, x = max, xend = min)) +
  geom_dumbbell(
    color = "darkgray",
    size = 1,
    size_x = 3,
    size_xend = 3,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  geom_dumbbell(
    data = gender_gap_data_internet %>% filter(country == "Overall mean"),
    aes(y = fct_relevel(country, "Overall mean"), x = max, xend = min),
    color = "darkgrey",
    size = 1.5,
    size_x = 4,
    size_xend = 4,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  labs(
    x = "Gender Gap Index, Internet",
    y = "",
    title = "Subnational Spread"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.y = element_text(size = 10, face = "bold"),
        plot.title = element_text(size = 13)
  ) + 
  scale_y_discrete(
    labels = c("Overall mean" = expression(bold("Overall mean")), "Other Country" = "Other Country"),
  ) + 
  geom_text(aes(x = (max + min) / 2-0.05, y = country, label = round(max-min, 2)), 
            vjust = -0.5, hjust = 0, size = 3.5, color = "black") +
  xlim(0, 1) +
  scale_x_continuous(
  limits = c(0, 1.02),
  breaks = c(0, 0.25, 0.5, 0.75, 1),
  labels = c(
    "0", "0.25", "0.5", "0.75", "1.00 \n(parity)"))

## Gender Gap Plot, mobile 

gender_gap_mobile <- gender_gap_data_mobile %>%
  mutate(diff = max - min) %>%
  mutate(
    country = fct_relevel(country, "Overall mean"), # Move "Overall mean" to the top
    country = fct_reorder(country, diff, .desc = FALSE) # Order by descending difference
  ) %>%
  ggplot(aes(y = country, x = max, xend = min)) +
  geom_dumbbell(
    color = "darkgray",
    size = 1,
    size_x = 3,
    size_xend = 3,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  geom_dumbbell(
    data = gender_gap_data_mobile %>% filter(country == "Overall mean"),
    aes(y = fct_relevel(country, "Overall mean"), x = max, xend = min),
    color = "darkgrey",
    size = 1.5,
    size_x = 4,
    size_xend = 4,
    colour_x = "#F69541",
    colour_xend = "#699DC6"
  ) +
  labs(
    x = "Gender Gap Index, Mobile",
    y = "",
    title = "Subnational Spread"
  ) +
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.y = element_text(size = 10, face = "bold"),
    plot.title = element_text(size = 13)
  ) + 
  scale_y_discrete(
    labels = c("Overall mean" = expression(bold("Overall mean")), "Other Country" = "Other Country"),
  ) + 
  geom_text(aes(x = (max + min) / 2-0.05, y = country, label = round(max-min, 2)), 
            vjust = -0.5, hjust = 0, size = 3.5, color = "black") +
  scale_x_continuous(
  limits = c(0, 1.02),
  breaks = c(0, 0.25, 0.5, 0.75, 1),
  labels = c(
    "0", "0.25", "0.5", "0.75", "1.00 \n(parity)"))


# Dummy plot for shared legend
legend_plot <- ggplot(data.frame(x = c(1, 2), y = c(1, 2), group = c("Lowest Gender Gap Index (Least Equal)", "Highest Gender Gap Index (Most Equal)")), 
                      aes(x = x, y = y, color = group)) +
  geom_point(size = 4) +
  scale_color_manual(
    values = c(
      "Lowest Gender Gap Index (Least Equal)" = "#699DC6",
      "Highest Gender Gap Index (Most Equal)" = "#F69541"
    ),
    breaks = c(
      "Lowest Gender Gap Index (Least Equal)",  # blue first
      "Highest Gender Gap Index (Most Equal)"   # orange second
    )
  ) + 
  guides(color = guide_legend(title = NULL)) +
  theme_void() +
  theme(legend.position = "bottom", legend.text = element_text(size = 10))

## Extract Legend 
legend <- ggpubr::get_legend(legend_plot)

## Combine into one plot 
combined_plot <- plot_grid(
  gender_gap_mobile, 
  gender_gap_internet, 
  labels = "AUTO", 
  ncol = 2
)

## Add on legend 
combined_plot_gender_disparities <- plot_grid(combined_plot, legend, ncol = 1, rel_heights = c(1, 0.1))

## Save plot 
ggsave(plot = combined_plot_gender_disparities, filename = here("figures", "fig5_disparities_plot.pdf"), width = 10, height = 5)
combined_plot_gender_disparities
```

```{r}
# Prepare Internet data for dumbbell plot
internet_data <- abs_diff_results %>%
  filter(str_detect(outcome, "internet")) %>%
  mutate(
    gender = ifelse(str_detect(outcome, "women"), "Women", "Men"),
    country = countrycode(gid_0, origin = "iso3c", destination = "country.name")
  ) %>%
  mutate(country = case_when(
    country == "Congo - Brazzaville" ~ "Republic of the Congo", 
    country == "Congo - Kinshasa" ~ "Democratic Republic of the Congo",
    TRUE ~ country
  )) %>% 
  group_by(outcome) %>%
  slice_max(diff, n = 10, with_ties = FALSE)  %>% # Select top 10 countries by gender
  bind_rows(abs_diff_results_mean) %>% 
  filter(str_detect(outcome, "internet")) 
```
