==> code/12_dhs_table.Rmd <==
---
title: "Additional appendix tables"
author: Casey Breen
---

Summary: This script reads in DHS data and creates a table

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```


```{r}
master_file <- read_csv(here("data", "master_data_file_march12_with_national.csv")) 
```

```{r}
# Define the list of country names in lowercase
countries <- c("angola", 
               "armenia",
               "benin",
               "burkina_faso", 
               "burundi",
               "cambodia", 
               "cameroon", 
               "cote_divoire",
               "ethiopia", 
               "gabon",
               "gambia",
               "guinea",
               "haiti", 
               "india", 
               "kenya",
               "liberia",
               "madagascar", 
               "malawi", 
               "mali",
               "mauritania", 
               "mozambique", 
               "nepal", 
               "nigeria", 
               "pakistan",
               "rwanda",
               "senegal", 
               "sierra_leone", 
               "south_africa", 
               "tanzania", 
               "timor-leste", 
               "uganda",
               "zambia",
               "zimbabwe")
 
# Define a function to check if a folder name contains any country names
folder_contains_country <- function(folder_name, countries) {
  any(sapply(countries, function(country) grepl(country, folder_name, ignore.case = TRUE)))
}

# List all folders in the specified paths
wave_paths <- c(
  "/Users/caseybreen/Library/CloudStorage/Dropbox-NuffieldCollege/Casey Breen/dgg_research/national/data_refresh/dhs/wave7/",
  "/Users/caseybreen/Library/CloudStorage/Dropbox-NuffieldCollege/Casey Breen/dgg_research/national/data_refresh/dhs/wave8/")


senegal_wom <- c("/Users/caseybreen/Library/CloudStorage/Dropbox-NuffieldCollege/Casey Breen/dgg_research/subnational/data_processing/dhs_survey_gps/Senegal/SN_2019_CONTINUOUSDHS_09192023_1943_198343/SNIR8BDT/SNIR8BFL.DTA") 


senegal_men <- c("/Users/caseybreen/Library/CloudStorage/Dropbox-NuffieldCollege/Casey Breen/dgg_research/subnational/data_processing/dhs_survey_gps/Senegal/SN_2019_CONTINUOUSDHS_09192023_1943_198343/SNMR8BDT/SNMR8BFL.DTA")

# Retrieve all folder paths
all_folders <- unlist(lapply(wave_paths, list.dirs, full.names = TRUE, recursive = FALSE))

# Filter folders containing any of the country names and exclude certain folders
selected_folders <- Filter(function(folder) folder_contains_country(folder, countries), all_folders)
selected_folders <- selected_folders[!grepl("cambodia_2014|india_2015|kenya_2014|rwanda_2014|rwanda_2014|nepal_2016|tanzania_2015", selected_folders, ignore.case = TRUE)]

# Initialize lists to store data
data_list_wom <- list()
data_list_men <- list()

# Loop through selected folders to read women's and men's survey files
for (folder in selected_folders) {
  # Women's files with 'IR' in the title
  files_wom <- list.files(path = folder, pattern = "IR", full.names = TRUE)
  files_wom <- c(files_wom, senegal_wom)
  for (file in files_wom) {
    data_list_wom[[file]] <- read_dta(file, col_select = c("v000", "v001", "v007", "v101", "v005", "v017", "v169a", "v171a")) %>% 
      mutate(v101 = as.numeric(v101)) %>% 
      mutate(path = file)
  }
  
  # Men's files with 'MR7' or 'MR8' in the title
  files_men <- list.files(path = folder, pattern = "MR7|MR8", full.names = TRUE)
  files_men <- c(senegal_men, files_men)
  for (file in files_men) {
    data_list_men[[file]] <- read_dta(file, col_select = c("mv000", "mv001", "mv007", "mv101", "mv005", "mv169a", "mv171a")) %>% 
      mutate(mv101 = as.numeric(mv101))  %>% 
      mutate(path = file)
  }
  cat(file) 
}

# Combine all women's and men's data into single data frames
dhs_surveys_wom <- bind_rows(data_list_wom)
dhs_surveys_men <- bind_rows(data_list_men)

```

```{r}
surveys_women <- dhs_surveys_wom %>% 
  group_by(v000) %>% 
  summarize(n_women = n(),
            start_year = min(v007),
            end_year = max(v007)) %>% 
  mutate(country_code = substring(v000, 1, 2)) %>% 
  dplyr::select(country_code, start_year, end_year, n_women)

surveys_men <- dhs_surveys_men %>% 
  group_by(mv000) %>% 
  summarize(n_men = n()) %>% 
  mutate(country_code = substring(mv000, 1, 2)) %>% 
  dplyr::select(country_code, n_men)


joint_files <- left_join(surveys_women, surveys_men , by = "country_code") %>% 
    mutate(country = countrycode::countrycode(country_code, origin = "iso2c", destination = "country.name")) %>%
    mutate(country = case_when(country_code == "BU" ~ "Burundi",
                               country_code == "IA" ~ "India",
                               country_code == "MD" ~ "Madagascar",
                               country_code ==  "LB" ~ "Liberia",
                               TRUE ~ country)) %>% 
  mutate(country_code = countrycode::countrycode(country_code, origin = "iso2c", destination = "iso3c"))

table <- joint_files %>% 
  dplyr::select(country, country_code, start_year, end_year, n_women, n_men) %>% 
  arrange(country) %>% 
  mutate(start_year = case_when(country == "Nepal" ~ 2022, TRUE ~ start_year)) %>% 
  mutate(end_year = case_when(country == "Nepal" ~ 2022, TRUE ~ end_year))

print(xtable(table, digits = 0), include.rownames=FALSE, file = here("figures", "dhs_table.tex"))

table %>% summarize(wom = sum(n_women), men = sum(n_men)) %>% 
  mutate(total = wom + men)
```


```{r}
master_file_full <- read_csv(here("data", "master_data_file_march12_with_national.csv"), 
                       locale = locale(encoding = "ISO-8859-1"),
                       show_col_types = FALSE) %>%
  janitor::clean_names() %>% 
  mutate(continent = case_when(continent == "Africa" ~ 1,
                               TRUE ~ 0)) %>% 
  filter(!(country == "Bangladesh" & dhsyear == "2014")) %>% 
  filter(!(country == "India" & dhsyear == "2014")) %>% 
  filter(!(country == "Cambodia" & dhsyear == "2014")) %>%
  filter(!(country == "Nepal" & dhsyear == "2016")) %>%
  filter(!(country == "Tanzania" & dhsyear == "2016")) %>%
  filter(!(country == "Philippines" & dhsyear == "2017")) %>%
  mutate(pop_all_age_total = pop_all_age_female + pop_all_age_male) %>%
  filter(!is.na(fb_pntr_18p_female)) %>% 
  mutate(
    nl_mean_zscore,
    pop_density,
    pop_all_age_female,
    pop_all_age_male,
    subnational_gdi,
    subnational_hdi_males,
    subnational_hdi_females,
    educational_index_females,
    educational_index_males,
    income_index_females,
    income_index_males,
    internet_speed,
    continent ,
    hdi,
    gdi)
```


```{r}
## columns 
master_file_select <- master_file %>% 
  dplyr::select(gid_0, country, dhs_vars, off_feats, fb_feats)

result_list <- list()
i = 1
for (col in colnames(master_file_select %>% dplyr::select(-country)) ) {
  
  total_country <- master_file_select %>% 
    filter(!is.na(!!sym(col))) %>% 
    distinct(country) %>% 
    count() %>% 
    pull()
  
  total_subnational <- master_file_select %>% 
    filter(!is.na(!!sym(col))) %>% 
    count() %>% 
    pull()
  
  total_subnational_metrics <- master_file_select %>% 
    filter(!is.na(!!sym(col))) %>% 
    summarize(min = min(!!sym(col)),
              median = median(!!sym(col)), 
              mean = mean(!!sym(col)), 
              max = max(!!sym(col)))
  
  df <- tibble(col, total_country, total_subnational) #%>% 
   # bind_cols(total_subnational_metrics)
  
  result_list[[i]] <- df 
  
  i = i + 1
}

table_variables <- bind_rows(result_list) 

variable_table <- table_variables %>% 
  mutate(across(where(is.numeric), ~round(., 1)))



print(xtable(variable_table, digits = 0), include.rownames=FALSE, file = here("figures", "variable_table.tex"))
```
