==> code/14_trends.Rmd <==
---
title: "Make trends figure"
author: Casey Breen
---

Summary: Create trend figure for main paper 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```


```{r}
predictions <- read_csv(here("out", "predictions_with_uncertainty.csv"))

predictions <- predictions %>% 
  mutate(date = as.Date(paste0(date, "-01")))
```




```{r}
# Filter for Guatemala
data_filter <- predictions %>%
    filter(outcome == "internet_fm_ratio") %>% 
  filter(gid_0 == "AFG")

# Calculate median line
median_data <- data_filter %>%
  group_by(date) %>%
  summarize(median_predicted = median(predicted, na.rm = TRUE))

# Plot with median as dark line and others as clear grey
afghanistan_plot <- ggplot() +
  geom_line(data = data_filter, aes(x = date, y = predicted, group = gid_1),
            color = "grey80", size = 0.5) +
  geom_line(data = median_data, aes(x = date, y = median_predicted),
            color = "black", size = 1) +
  theme_cowplot() +
  labs(
    title = "Afghanistan",
    x = "Date",
    y = "Gender Gap Index (Internet)"
  ) + 
    background_grid() + 
  scale_y_continuous(
    limits = c(0, 1),
    breaks = c(0.00, 0.25, 0.50, 0.75, 1.00),
    labels = c("0.00", "0.25", "0.50", "0.75", "(Parity) 1.00"))
```


```{r}
# Filter for Guatemala
data_filter <- predictions %>%
    filter(outcome == "internet_fm_ratio") %>% 
  filter(gid_0 == "NGA")

# Calculate median line
median_data <- data_filter %>%
  group_by(date) %>%
  summarize(median_predicted = median(predicted, na.rm = TRUE))

# Plot with median as dark line and others as clear grey
nigeria_plot <- ggplot() +
  geom_line(data = data_filter, aes(x = date, y = predicted, group = gid_1),
            color = "grey80", size = 0.5) +
  geom_line(data = median_data, aes(x = date, y = median_predicted),
            color = "black", size = 1) +
  theme_cowplot() +
  background_grid() + 
  labs(
    title = "Nigeria",
    x = "Date",
    y = "Gender Gap Index (Internet)"
  ) + 
  scale_y_continuous(
    limits = c(0, 1),
    breaks = c(0.00, 0.25, 0.50, 0.75, 1.00),
    labels = c("0.00", "0.25", "0.50", "0.75", "(Parity) 1.00"))


nigeria_plot
```


```{r}
# Filter for Guatemala
data_filter <- predictions %>%
    filter(outcome == "internet_fm_ratio") %>% 
  filter(gid_0 == "IND")

# Calculate median line
median_data <- data_filter %>%
  group_by(date) %>%
  summarize(median_predicted = median(predicted, na.rm = TRUE))

# Plot with median as dark line and others as clear grey
india_plot <- ggplot() +
  geom_line(data = data_filter, aes(x = date, y = predicted, group = gid_1),
            color = "grey80", size = 0.5) +
  geom_line(data = median_data, aes(x = date, y = median_predicted),
            color = "black", size = 1) +
  theme_cowplot() +
    background_grid() + 
  labs(
    title = "India",
    x = "Date",
    y = "Gender Gap Index (Internet)"
  ) + 
  scale_y_continuous(
    limits = c(0, 1),
    breaks = c(0.00, 0.25, 0.50, 0.75, 1.00),
    labels = c("0.00", "0.25", "0.50", "0.75", "(Parity) 1.00"))
```

```{r}
trend_plot <- plot_grid(india_plot, nigeria_plot, afghanistan_plot, nrow = 1, labels = "AUTO")
```


## internet figure 


```{r}
# Filter for the outcome of interest (e.g., "internet_women")
filtered_data <- predictions %>%
  filter(outcome == "internet_women")

# Calculate Gini coefficient for each country and year
gini_results <- filtered_data %>%
  group_by(gid_0, year) %>%
  summarise(
    gini = ineq(predicted, type = "Gini"),  # Compute Gini coefficient for 'predicted' values
    .groups = "drop"
  )
```



```{r}
gini_results <- gini_results %>%
  mutate(continent = countrycode(sourcevar = gid_0, origin = "iso3c", destination = "continent")) %>% 
  mutate(continent = case_when(
    gid_0 %in% c("IND", "PAK", "BGD", "LKA", "NPL", "BTN", "MDV", "AFG") ~ "South Asia",
    TRUE ~ continent)) %>% 
  mutate(continent =  case_when(
    continent == "Asia" ~ "Asia (other)",
    TRUE ~ continent))

gini_plot <- gini_results %>% 
  filter(continent %in% c("Africa", "Americas", "Asia (other)", "South Asia")) %>% 
  group_by(year, continent) %>% 
  filter(!is.na(gini)) %>% 
  summarize(gini = mean(gini, na.rm = TRUE)) %>% 
  ggplot(aes(x = year, y = gini, color = continent, shape = continent)) + 
  geom_line() + 
  geom_point(size = 4, fill = "white", stroke = 1.5) +  # fill only applies to shape 21+
  theme_cowplot() + 
  scale_y_continuous(
  limits = c(0, 0.4),
  breaks = c(0, 0.1, 0.2, 0.3, 0.4),
  labels = c("(Perfect Equality) 0.0", "0.1", "0.2", "0.3", "0.4")
) + 
  theme(legend.position = "bottom") + 
  scale_x_continuous(breaks = seq(min(gini_results$year, na.rm = TRUE), 
                                  max(gini_results$year, na.rm = TRUE), by = 1)) + 
  scale_color_manual(values = cud) + 
  scale_shape_manual(values = c(21, 22, 23, 24)) +  # Assign different shapes to continents
  background_grid() + 
  labs(x = "Date", y = "Gini Index")
```

```{r}
trend_plot_combined <- cowplot::plot_grid(trend_plot, gini_plot, ncol = 1, labels = c("", "D"))

ggsave(plot = trend_plot_combined , filename = here("figures", "fig7_trend_plot_combined.pdf"), height = 8, width = 11)
```




```{r}
# Filter for Guatemala
data_filter <- predictions %>%
  filter(gid_0 == "AFG") %>% 
  filter(outcome %in% c("internet_fm_ratio", "internet_men", "internet_women"))

# Calculate median line
median_data <- data_filter %>%
  group_by(date, outcome) %>%
  summarize(median_predicted = median(predicted, na.rm = TRUE))

# Plot with median as dark line and others as clear grey
afghanistan_plot <- ggplot() +
  geom_line(data = data_filter, aes(x = date, y = predicted, group = gid_1),
            color = "grey80", size = 0.5) +
  geom_line(data = median_data, aes(x = date, y = median_predicted),
            color = "black", size = 1) +
  theme_cowplot() +
  labs(
    title = "Afghanistan",
    x = "Date",
  ) + 
  background_grid() + 
  facet_wrap(~outcome) + 
  ylim(0, 0.5)
```


```{r}
# Filter for Guatemala
data_filter <- predictions %>%
  filter(gid_0 == "AFG") %>% 
  filter(outcome %in% c("mobile_fm_ratio", "mobile_men", "mobile_women"))

# Calculate median line
median_data <- data_filter %>%
  group_by(date, outcome) %>%
  summarize(median_predicted = median(predicted, na.rm = TRUE))

# Plot with median as dark line and others as clear grey
afghanistan_plot <- ggplot() +
  geom_line(data = data_filter, aes(x = date, y = predicted, group = gid_1),
            color = "grey80", size = 0.5) +
  geom_line(data = median_data, aes(x = date, y = median_predicted),
            color = "black", size = 1) +
  theme_cowplot() +
  labs(
    title = "Afghanistan",
    x = "Date",
  ) + 
  background_grid() + 
  facet_wrap(~outcome) + 
  ylim(0, 1)
```


```{r}
# Filter for Guatemala
data_filter <- predictions %>%
    filter(outcome == "internet_fm_ratio") %>% 
  filter(gid_0 == "AFG")

# Calculate median line
median_data <- data_filter %>%
  group_by(date) %>%
  summarize(median_predicted = median(predicted, na.rm = TRUE))

# Plot with median as dark line and others as clear grey
afghanistan_plot <- ggplot() +
  geom_line(data = median_data, aes(x = date, y = median_predicted),
            color = "black", size = 1) +
    geom_point(data = median_data, aes(x = date, y = median_predicted),
            color = "black", fill = "white", shape = 21, size = 3) + 
  theme_cowplot() +
  labs(
    title = "Afghanistan",
    x = "Date",
    y = "Gender Gap (Internet)"
  ) + 
 background_grid() + 
  ylim(0, .4) + 
  geom_vline(xintercept = as.Date("2021-08-01"), linetype = "dashed", color = "firebrick") + 
 # Improved text annotation (moved slightly to the right)
  annotate("text", x = as.Date("2022-01-01"), 
           y = 0.22, label = "Taliban Resurgence", 
           size = 4, colour = "firebrick", fontface = "bold", hjust = 0) +

  # Left-pointing arrow segment to connect text to the event
  annotate("segment", 
           x = as.Date("2021-12-20"), xend = as.Date("2021-09-01"), 
           y = 0.22, yend = 0.22,
           arrow = arrow(length = unit(0.2, "cm")), 
           color = "firebrick", size = 0.8)

```
