==> code/16_validation_of_uncertainty.Rmd <==
---
title: "Predicted error"
author: Casey Breen
---

## Summary 

Compare predicted error vs. real LSMS error 

```{r}
## libary packages + custom functions 
library(here)
source(here("code", "helpers.R"))
```

```{r}
## Read in data 
ground_truth_lsms <- read_csv(here("data", "subnational_ground_truth.csv")) %>% 
  filter(survey_type == "lsms")

## read in predictions 
predictions <- read_csv(here("out", "predictions_with_uncertainty.csv"))
```




```{r}
## estimates longs 
lsms_estimates_long <- ground_truth_lsms %>% 
 filter(sample_size > 150) 

lsms_estimates_long <- lsms_estimates_long %>% 
  filter(!(gid_0 == "NGA" & outcome %in% c("mobile_women", "mobile_men", "mobile_fm_ratio")))
```


```{r}
# Joining Data
lsms_data <- predictions %>% 
  select(gid_1, predicted, predicted_error, outcome, year) %>% 
  inner_join(lsms_estimates_long, by = c("year", "gid_1", "outcome")) %>% 
  filter(!is.na(observed) & !is.na(predicted)) #%>%  filter(iso3 != "NER")


lsms_data <-  lsms_data %>% 
   mutate(outcome_clean = case_when(
    grepl("mobile_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_women", outcome) ~ "Mobile (Women)",
    grepl("internet_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("internet_men", outcome) ~ "Internet (Men)",
    grepl("internet_women", outcome) ~ "Internet (Women)",
    TRUE ~ outcome
  )) %>% mutate(outcome_clean = factor(outcome_clean, 
                                levels = c("Internet (Women)", "Internet (Men)", "Internet (Ratio)",
                                           "Mobile (Women)", "Mobile (Men)", "Mobile (Ratio)")))

lsms_error_plot <-  lsms_data %>% 
  group_by(outcome_clean) %>% 
  mutate(error = predicted - observed) %>% 
  summarize(`Observed error (LSMS)` = mean(abs(error)), `Predicted error` = mean(predicted_error)) %>% 
  pivot_longer(-outcome_clean) %>% 
  ggplot(aes(x = outcome_clean, y = value, fill = name)) + 
  geom_col(position = "dodge", alpha = 0.5, color = "black") + 
  scale_fill_viridis_d() + 
  theme_cowplot() + 
  theme(legend.position = "bottom") + 
  labs(x = "",
       y = "Mean Error",
       fill = "")

## lsms error plots 
ggsave(plot = lsms_error_plot, filename = here("figures", "lsms_error_plot.png"), width = 10, height = 5)
```
