==> code/18_mics_comparison.Rmd <==
---
title: "MICS Surveys"
author: Casey Breen
---

Summary: Calculate subnational estimates for MICS surveys 


```{r}
## Library packages + custom functions 
library(here)
source(here("code", "helpers.R"))

## Read in predictions
predictions <- read_csv(here("out", "predictions_with_uncertainty.csv")) 

## Read in data file with all subnational estimates 
ground_truth_mics <- read_csv(here("data", "subnational_ground_truth.csv")) %>% 
  filter(survey_type == "mics")

ground_truth_mics_wide <- read_csv(here("data", "subnational_ground_truth_wide.csv")) %>% 
  filter(survey_type == "mics")
```
```{r}
## Filter to only include countries with enough data  
ground_truth_mics <- ground_truth_mics %>% 
  filter(sample_size >= 150) 
```


```{r}
## create table 
mics_table <- ground_truth_mics_wide %>%
 # filter(outcome %in% c("mobile_women", "mobile_men")) %>%
  group_by(gid_0, year) %>%
  summarize(female_sample_internet = sum(internet_n_sample_f, na.rm = TRUE),
            female_sample_mobile = sum(mobile_n_sample_f, na.rm = TRUE),
            male_sample_internet = sum(internet_n_sample_m, na.rm = TRUE), 
            male_sample_mobile = sum(mobile_n_sample_m, na.rm = TRUE)) %>%
  rename(country_code = gid_0) %>% 
  mutate(country = countrycode::countrycode(country_code, origin = "iso3c", destination = "country.name")) %>% 
    select(country, country_code, year, female_sample_internet, male_sample_internet, female_sample_mobile, male_sample_mobile)

# Add total row
total_row <- mics_table %>%
  ungroup() %>% 
  summarize(
    country = "Total",
    country_code = "",
    year = NA,
    female_sample_internet = sum(female_sample_internet, na.rm = TRUE),
    male_sample_internet = sum(male_sample_internet, na.rm = TRUE),
    female_sample_mobile = sum(female_sample_mobile, na.rm = TRUE),
    male_sample_mobile = sum(male_sample_mobile, na.rm = TRUE)
  )

# Bind total row to table
mics_table <- bind_rows(mics_table, total_row)

## print table 
print(xtable(mics_table, digits = 0), include.rownames=FALSE, file = here("figures", "mics_comparison_table.tex"))
```



```{r}
## create table 
mics_table <- ground_truth_mics %>%
  filter(outcome %in% c("mobile_women", "mobile_men")) %>%
  mutate(gender = if_else(outcome == "mobile_women", "female", "male")) %>%
  group_by(gid_0, year, gender) %>%
  summarize(total_sample = sum(sample_size, na.rm = TRUE), .groups = "drop") %>%
  pivot_wider(
    names_from = gender,
    values_from = total_sample,
    names_glue = "{gender}_sample"
  ) %>%
  rename(country_code = gid_0) %>% 
  mutate(country = countrycode::countrycode(country_code, origin = "iso3c", destination = "country.name")) %>% 
  select(country, country_code, year, female_sample, male_sample)

# Add total row
total_row <- mics_table %>%
  summarize(
    country = "Total",
    country_code = "",
    year = NA,
    female_sample = sum(female_sample, na.rm = TRUE),
    male_sample = sum(male_sample, na.rm = TRUE)
  )

mics_table <- bind_rows(mics_table, total_row)

## print table 
print(xtable(mics_table, digits = 0), include.rownames=FALSE, file = here("figures", "mics_comparison_table.tex"))
```



```{r}
## Prepare df 
subnational_predictions_comparison_df <- ground_truth_mics %>% 
  left_join(predictions, by = join_by(gid_0, gid_1, year, outcome)) %>% 
  distinct() %>% 
   mutate(outcome_clean = case_when(
    grepl("mobile_fm_ratio", outcome) ~ "Mobile (Ratio)",
    grepl("mobile_men", outcome) ~ "Mobile (Men)",
    grepl("mobile_women", outcome) ~ "Mobile (Women)",
    grepl("internet_fm_ratio", outcome) ~ "Internet (Ratio)",
    grepl("internet_men", outcome) ~ "Internet (Men)",
    grepl("internet_women", outcome) ~ "Internet (Women)",
    TRUE ~ outcome
  ))

# Define the order of the levels explicitly
subnational_predictions_comparison_df <- subnational_predictions_comparison_df %>% 
  mutate(outcome_clean = factor(outcome_clean, 
                                levels = c("Internet (Women)", "Internet (Men)", "Internet (Ratio)",
                                           "Mobile (Women)", "Mobile (Men)", "Mobile (Ratio)")))

# Calculate R-squared and correlation (r) by outcome
r_squared_results <- subnational_predictions_comparison_df %>%
  filter(!is.na(observed) & !is.na(predicted) & !is.infinite(observed)) %>%  #%>%  filter(iso3 != "NER")
  group_by(outcome_clean) %>%
  summarize(
    r = cor(observed, predicted, use = "complete.obs"),  # Calculate correlation
    r2 = compute_r2(observed, predicted),  # Calculate R-squared
    mae = mean(abs(observed - predicted)), 
    .groups = "drop"
  )

## Subnational predictions comparison 
mics_comparison_fig <- subnational_predictions_comparison_df %>% 
  left_join(r_squared_results) %>% 
  ggplot(aes(x = observed, y = predicted, color = year)) + 
  geom_point(size = 2, alpha = 0.85) +
  geom_abline(color = "grey60", linetype = "dashed") +
  xlim(0, 1) + 
  ylim(0, 1) + 
  labs(
    x = "Observed, MICS (Admin-1)",
    y = "Predicted (Admin-1)",
    color = "MICS Survey Year"
  ) +
  geom_label(
    aes(
      label = paste0("atop(italic(r)==", round(r, 2), ",~italic(MAE)==", round(mae, 2), ")")
    ),
    parse = TRUE,
    x = 0.20,
    y = 0.9,
    size = 3.2,
    inherit.aes = FALSE
  ) +
  facet_wrap(~ outcome_clean, ncol = 3) +
  theme_cowplot() +
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(1.5, "lines"),
    aspect.ratio = .9,
    legend.position = "bottom",
    legend.justification = "center",  # centers the whole legend box
    legend.title = element_text(size = 13, hjust = 0.5, vjust = 1, margin = margin(b = 4)),
    legend.text = element_text(size = 11)
  ) +
  guides(
    color = guide_colorbar(
      title.position = "top",
      title.hjust = 0.5,
      barwidth = unit(10, "cm"),
      barheight = unit(0.5, "cm")))

## Save plot 
ggsave(plot = mics_comparison_fig, filename = here("figures", "mics_comparison_subnational.png"), height = 7, width = 10)
```


```{r}
## Subnational predictions 
subnational_predictions_comparison_df_select <- subnational_predictions_comparison_df %>% 
  filter(gid_0 %in% c("SUR", "GUY", "JAM"))

# Calculate R-squared and correlation (r) by outcome
r_squared_results <- subnational_predictions_comparison_df_select %>% 
  group_by(outcome) %>%
  summarize(
    r = cor(predicted, observed, use = "complete.obs"),  # Calculate correlation
    r2 = compute_r2(predicted, observed),  # Calculate R-squared
    mae = mean(abs(predicted - observed)), 
    .groups = "drop"
  )

mics_survey <- subnational_predictions_comparison_df_select %>% 
  left_join(r_squared_results) %>% 
  mutate(survey = case_when(
    gid_0 == "SUR" ~ "Suriname (2018)",
    gid_0 == "GUY" ~ "Guyana (2019)", 
    gid_0 == "JAM" ~ "Jamaica (2022)"
  )) %>% 
  ggplot(aes(x = observed, y = predicted, color = survey, shape = survey)) + 
  geom_point(size = 2, alpha = 0.85) +
  theme_cowplot() + 
  geom_abline(color = "grey60", linetype = "dashed") +
  ylim(0, 1.1) + 
  geom_label(
    aes(label = paste("atop(italic(r) == ", round(r, 2), ", italic(MAE) == ", round(mae, 2), ")")),
    x = 0.25,  
    y = 0.9, 
    size = 3.5,
    parse = TRUE,
    inherit.aes = FALSE
  ) +
  xlim(0, 1.1) + 
  theme(legend.position = "bottom") + 
    facet_wrap(~outcome_clean) + 
  labs(y = "Predicted (Admin-1)",
       x = "Observed, MICS (Admin-1)") + 
  theme(
    strip.background = element_rect(fill = "white", color = "black", linewidth = 0.8),
    strip.text = element_text(size = 11, face = "bold", color = "black"),
    panel.border = element_rect(color = "black", fill = NA, linewidth = 0.5),
    panel.background = element_rect(fill = "#f9f9f9", color = NA),
    panel.grid = element_line(color = "grey95"),
    panel.spacing = unit(1.5, "lines"),
    aspect.ratio = .9,
    legend.position = "bottom",
    legend.justification = "center",  # centers the whole legend box
    legend.title = element_text(size = 13, hjust = 0.5, vjust = 1, margin = margin(b = 4)),
    legend.text = element_text(size = 11)
  ) + 
  scale_color_viridis_d()

## save plot 
ggsave(plot = mics_survey, filename = here("figures", "mics_comparison_subnational_latin_america.png"), width = 10, height = 7, bg = "white")
```
